+++ test/TestApp/Program.cs
@@ -83,8 +83,6 @@ namespace TestApp
             Stopwatch watch = new Stopwatch();
             watch.Start();
 
-            new OperatorTests().GaussianIsBetweenTest2();
-
             if (false)
             {
                 // Run all tests (need to run in 64-bit else OutOfMemory due to loading many DLLs)
