+++ test/Tests/Operators/OperatorTests.cs
@@ -18,19 +18,13 @@ namespace Microsoft.ML.Probabilistic.Tests
 
     using Utilities;
 
-    using Assert = Microsoft.ML.Probabilistic.Tests.AssertHelper;
+    using Assert = AssertHelper;
     using GaussianArray = DistributionStructArray<Gaussian, double>;
     using GaussianArray2D = DistributionStructArray2D<Gaussian, double>;
 
 
     public class OperatorTests
     {
-        [Fact]
-        public void GammaPowerProductOp_LaplaceTest()
-        {
-            GammaPowerProductOp_Laplace.ProductAverageConditional(GammaPower.Uniform(-1), new GammaPower(22, 0.04762, -1), new GammaPower(4.984e+10, 9.32e-120, -1), new Gamma(4.984e+10, 9.32e-120), GammaPower.Uniform(-1));
-        }
-
         [Fact]
         public void GaussianPlusOpTest()
         {
@@ -228,1186 +222,6 @@ namespace Microsoft.ML.Probabilistic.Tests
                     //Trace.WriteLine($"precision={precision} yv={yv}: {result0.Rate} {result.Rate}");
                 }
             }
-        }
-
-        // Test inference on a model where precision is scaled.
-        internal void GammaProductTest()
-        {
-            for (int i = 0; i <= 20; i++)
-            {
-                double minutesPlayed = System.Math.Pow(0.1, i);
-                if (i == 20) minutesPlayed = 0;
-                var EventCountMean_F = Gaussian.PointMass(0);
-                var EventCountPrecision_F = GammaRatioOp.RatioAverageConditional(Gamma.PointMass(1), minutesPlayed);
-                var realCount_F = GaussianOp_PointPrecision.SampleAverageConditional(EventCountMean_F, EventCountPrecision_F);
-                var realCount_use_B = MaxGaussianOp.BAverageConditional(0.0, 0.0, realCount_F);
-                var EventCountPrecision_B = GaussianOp_PointPrecision.PrecisionAverageConditional(realCount_use_B, EventCountMean_F, EventCountPrecision_F);
-                var EventsPerMinutePrecision_B = GammaRatioOp.AAverageConditional(EventCountPrecision_B, minutesPlayed);
-                Console.WriteLine($"realCount_use_B = {realCount_use_B}, EventCountPrecision_B = {EventCountPrecision_B},  EventsPerMinutePrecision_B = {EventsPerMinutePrecision_B}");
-            }
-        }
-
-        [Fact]
-        public void VariablePointOp_RpropGammaTest()
-        {
-            using (TestUtils.TemporarilyUseMeanPointGamma)
-            {
-                var buffer = VariablePointOp_RpropGamma.Buffer0Init();
-                Gamma g = Gamma.FromShapeAndRate(0.6, 1);
-                Gamma to_marginal = Gamma.PointMass(1);
-                for (int i = 0; i < 1000; i++)
-                {
-                    buffer = VariablePointOp_RpropGamma.Buffer0(g, g, to_marginal, buffer);
-                    //Console.WriteLine(buffer.nextPoint);
-                    to_marginal = Gamma.Uniform();
-                }
-            }
-        }
-
-        [Fact]
-        public void ArrayFromVectorOpTest()
-        {
-            bool verbose = false;
-            PositiveDefiniteMatrix A = new PositiveDefiniteMatrix(new double[,] {
-                { 4.008640513161180,  1.303104352135630, - 2.696380025254830, - 2.728465435435790 },
-                { 1.303104352135630,  4.024136989099960, - 2.681070246787840, - 2.713155656968810 },
-                { -2.696380025254830, - 2.681070246787840, 4.136120496920130,  1.403451295855420 },
-                { -2.728465435435790, - 2.713155656968810, 1.403451295855420,  4.063123100392480 }
-            });
-            Vector arrayVariance = Vector.FromArray(1, 0, 3, 4);
-            Vector arrayMean = Vector.FromArray(1, 2, 3, 4);
-
-            VectorGaussianMoments vg = new VectorGaussianMoments(Vector.FromArray(6, 5, 4, 3), A);
-            Gaussian[] array = Util.ArrayInit(arrayMean.Count, i => new Gaussian(arrayMean[i], arrayVariance[i]));
-            var result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
-            Vector varianceExpected = Vector.FromArray(0.699231932932321, 0, 0.946948669226297, 0.926496676481940);
-            Vector varianceActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetVariance()));
-            if (verbose) Console.WriteLine($"variance = {varianceActual} should be {varianceExpected}");
-            Assert.True(varianceExpected.MaxDiff(varianceActual) < 1e-4);
-            Vector meanExpected = Vector.FromArray(2.640276200841019, 2.000000014880260, 6.527941507328482, 6.908179339051594);
-            Vector meanActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetMean()));
-            if (verbose) Console.WriteLine($"mean = {meanActual} should be {meanExpected}");
-            Assert.True(meanExpected.MaxDiff(meanActual) < 1e-4);
-
-            arrayVariance.SetTo(Vector.FromArray(1, double.PositiveInfinity, 3, 4));
-            array = Util.ArrayInit(arrayMean.Count, i => new Gaussian(arrayMean[i], arrayVariance[i]));
-            result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
-            varianceExpected = Vector.FromArray(0.703202829692760, 2.371534574978036, 1.416576208767429, 1.566923316764467);
-            varianceActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetVariance()));
-            if (verbose) Console.WriteLine($"variance = {varianceActual} should be {varianceExpected}");
-            Assert.True(varianceExpected.MaxDiff(varianceActual) < 1e-4);
-            meanExpected = Vector.FromArray(2.495869677904531, 5.528904975989440, 4.957577684071078, 5.074347889058121);
-            meanActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetMean()));
-            if (verbose) Console.WriteLine($"mean = {meanActual} should be {meanExpected}");
-            Assert.True(meanExpected.MaxDiff(meanActual) < 1e-4);
-
-            bool testImproper = false;
-            if (testImproper)
-            {
-                array = new Gaussian[]
-                {
-                new Gaussian(0.5069, 27.43),
-                new Gaussian(0.649, 38.94),
-                Gaussian.FromNatural(-0.5753, 0),
-                new Gaussian(0.1064, 10.15)
-                };
-                PositiveDefiniteMatrix B = new PositiveDefiniteMatrix(new double[,]
-                {
-                { 0.7158,  -0.1356, -0.2979, -0.2993 },
-                { -0.1356, 0.7294,  -0.2975, -0.2989 },
-                { -0.2979, -0.2975, 0.7625,  -0.1337 },
-                { -0.2993, -0.2989, -0.1337, 0.7203 },
-                });
-                vg = new VectorGaussianMoments(Vector.FromArray(0.3042, 0.4795, -0.2114, -0.5748), B);
-                result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
-                Gaussian[] resultExpected = new Gaussian[] {
-                    new Gaussian(0.4589, 0.707),
-                    new Gaussian(0.6338, 0.7204),
-                    new Gaussian(-0.2236, 0.7553),
-                    new Gaussian(-0.4982, 0.7148),
-                };
-                if (verbose) Console.WriteLine(StringUtil.ArrayToString(result));
-                for (int i = 0; i < result.Length; i++)
-                {
-                    Assert.True(resultExpected[i].MaxDiff(result[i]) < 1e-4);
-                }
-            }
-        }
-
-        internal void VectorFromArrayOp_SpeedTest()
-        {
-            int dim = 20;
-            int n = 100000;
-            VectorGaussian vector = new VectorGaussian(dim);
-            Matrix random = new Matrix(dim, dim);
-            for (int i = 0; i < dim; i++)
-            {
-                vector.MeanTimesPrecision[i] = Rand.Double();
-                for (int j = 0; j < dim; j++)
-                {
-                    random[i, j] = Rand.Double();
-                }
-            }
-            vector.Precision.SetToOuter(random);
-            Gaussian[] array = Util.ArrayInit(dim, i => Gaussian.FromMeanAndVariance(i, i));
-            Gaussian[] result = new Gaussian[dim];
-            for (int i = 0; i < n; i++)
-            {
-                VectorFromArrayOp.ArrayAverageConditional(vector, array, result);
-            }
-        }
-
-        [Fact]
-        public void VectorFromArrayOp_PointMassTest()
-        {
-            VectorFromArrayOp_HandlesPointMass(false);
-            VectorFromArrayOp_HandlesPointMass(true);
-        }
-
-        private void VectorFromArrayOp_HandlesPointMass(bool partial)
-        {
-            int dim = 2;
-            VectorGaussian to_vector = new VectorGaussian(dim);
-            VectorGaussian vector = VectorGaussian.FromNatural(Vector.FromArray(2.0, 3.0), new PositiveDefiniteMatrix(new double[,] { { 5.0, 1.0 }, { 1.0, 6.0 } }));
-            GaussianArray expected = null;
-            double lastError = double.PositiveInfinity;
-            for (int i = 0; i < 10; i++)
-            {
-                double variance = System.Math.Exp(-i);
-                if (i == 0)
-                    variance = 0;
-                Gaussian[] array = Util.ArrayInit(dim, j => new Gaussian(j, (j > 0 && partial) ? j : variance));
-                to_vector = VectorFromArrayOp.VectorAverageConditional(array, to_vector);
-                var to_array = new GaussianArray(dim);
-                to_array = VectorFromArrayOp.ArrayAverageConditional(vector, array, to_array);
-                if (i == 0)
-                    expected = to_array;
-                else
-                {
-                    double error = expected.MaxDiff(to_array);
-                    Assert.True(error < lastError);
-                    lastError = error;
-                }
-                ////Console.WriteLine(to_array);
-            }
-        }
-
-        [Fact]
-        public void GammaRatioOpTest()
-        {
-            Gamma ratio = new Gamma(2, 3);
-            double shape = 4;
-            Gamma A = new Gamma(shape, 1);
-            Gamma B = new Gamma(5, 6);
-            Gamma q = GammaFromShapeAndRateOp_Laplace.Q(ratio, shape, B);
-            Gamma bExpected = GammaFromShapeAndRateOp_Laplace.RateAverageConditional(ratio, shape, B, q);
-            Gamma rExpected = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(ratio, shape, B, q);
-            q = GammaRatioOp_Laplace.Q(ratio, A, B);
-            Gamma bActual = GammaRatioOp_Laplace.BAverageConditional(ratio, A, B, q);
-            Gamma rActual = GammaRatioOp_Laplace.RatioAverageConditional(ratio, A, B);
-            Console.WriteLine("b = {0} should be {1}", bActual, bExpected);
-            Console.WriteLine("ratio = {0} should be {1}", rActual, rExpected);
-            Assert.True(bExpected.MaxDiff(bActual) < 1e-4);
-            Assert.True(rExpected.MaxDiff(rActual) < 1e-4);
-        }
-
-        // Test that the operator behaves correctly for arguments with small variance
-        [Fact]
-        public void GammaFromShapeAndRateOpTest()
-        {
-            Assert.False(double.IsNaN(GammaFromShapeAndRateOp_Slow.SampleAverageConditional(Gamma.PointMass(0), 2.0, new Gamma(1, 1)).Rate));
-            Assert.False(double.IsNaN(GammaFromShapeAndRateOp_Slow.RateAverageConditional(new Gamma(1, 1), 2.0, Gamma.PointMass(0)).Rate));
-
-            Gamma sample, rate, result;
-            double prevDiff;
-            double shape = 3;
-            rate = Gamma.FromShapeAndRate(4, 5);
-            sample = Gamma.PointMass(2);
-            result = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-            Console.WriteLine("{0}: {1}", sample, result);
-            prevDiff = double.PositiveInfinity;
-            for (int i = 10; i < 50; i++)
-            {
-                double v = System.Math.Pow(0.1, i);
-                sample = Gamma.FromMeanAndVariance(2, v);
-                Gamma result2 = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-                double diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} (diff={2})", sample, result2, diff.ToString("g4"));
-                Assert.True(diff <= prevDiff || diff < 1e-10);
-                prevDiff = diff;
-            }
-        }
-        [Fact]
-        public void GammaFromShapeAndRateOpTest2()
-        {
-            Gamma sample, rate, result;
-            double prevDiff;
-            double shape = 3;
-            sample = Gamma.FromShapeAndRate(4, 5);
-            shape = 0.1;
-            rate = Gamma.FromShapeAndRate(0.1, 0.1);
-            result = GammaFromShapeAndRateOp_Slow.RateAverageConditional(sample, shape, rate);
-            Console.WriteLine(result);
-
-            shape = 3;
-            rate = Gamma.PointMass(2);
-            result = GammaFromShapeAndRateOp_Slow.RateAverageConditional(sample, shape, rate);
-            Console.WriteLine("{0}: {1}", rate, result);
-            prevDiff = double.PositiveInfinity;
-            for (int i = 10; i < 50; i++)
-            {
-                double v = System.Math.Pow(0.1, i);
-                rate = Gamma.FromMeanAndVariance(2, v);
-                Gamma result2 = GammaFromShapeAndRateOp_Slow.RateAverageConditional(sample, shape, rate);
-                double diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} (diff={2})", rate, result2, diff.ToString("g4"));
-                Assert.True(diff <= prevDiff || diff < 1e-10);
-                prevDiff = diff;
-            }
-        }
-        [Fact]
-        public void GammaFromShapeAndRateOpTest3()
-        {
-            Console.WriteLine(GammaFromShapeAndRateOp_Slow.RateAverageConditional(Gamma.FromShapeAndRate(2, 1), 1, Gamma.FromShapeAndRate(1.5, 0.5)));
-
-            Gamma sample = Gamma.FromShapeAndRate(2, 0);
-            Gamma rate = Gamma.FromShapeAndRate(4, 1);
-            double shape = 1;
-
-            Gamma to_sample2 = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-            double evExpected = GammaFromShapeAndRateOp_Slow.LogEvidenceRatio(sample, shape, rate, to_sample2);
-            Console.WriteLine("sample = {0} to_sample = {1} evidence = {2}", sample, to_sample2, evExpected);
-            for (int i = 40; i < 41; i++)
-            {
-                sample.Rate = System.Math.Pow(0.1, i);
-                Gamma to_sample = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-                double evActual = GammaFromShapeAndRateOp_Slow.LogEvidenceRatio(sample, shape, rate, to_sample);
-                Console.WriteLine("sample = {0} to_sample = {1} evidence = {2}", sample, to_sample, evActual);
-                Assert.True(to_sample2.MaxDiff(to_sample2) < 1e-4);
-                Assert.True(MMath.AbsDiff(evExpected, evActual) < 1e-4);
-            }
-        }
-        [Fact]
-        public void GammaFromShapeAndRateOpTest4()
-        {
-            Gamma sample = Gamma.FromShapeAndRate(2, 0);
-            Gamma rate = Gamma.FromShapeAndRate(4, 1);
-            double shape = 1;
-
-            Gamma rateExpected = GammaFromShapeAndRateOp_Slow.RateAverageConditional(sample, shape, rate);
-            Gamma q = GammaFromShapeAndRateOp_Laplace.Q(sample, shape, rate);
-            Gamma rateActual = GammaFromShapeAndRateOp_Laplace.RateAverageConditional(sample, shape, rate, q);
-            Assert.True(rateExpected.MaxDiff(rateActual) < 1e-4);
-
-            Gamma to_sample2 = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(sample, shape, rate, q);
-            double evExpected = GammaFromShapeAndRateOp_Laplace.LogEvidenceRatio(sample, shape, rate, to_sample2, q);
-            Console.WriteLine("sample = {0} to_sample = {1} evidence = {2}", sample, to_sample2, evExpected);
-            for (int i = 40; i < 41; i++)
-            {
-                sample.Rate = System.Math.Pow(0.1, i);
-                q = GammaFromShapeAndRateOp_Laplace.Q(sample, shape, rate);
-                Gamma to_sample = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(sample, shape, rate, q);
-                double evActual = GammaFromShapeAndRateOp_Laplace.LogEvidenceRatio(sample, shape, rate, to_sample, q);
-                Console.WriteLine("sample = {0} to_sample = {1} evidence = {2}", sample, to_sample, evActual);
-                Assert.True(to_sample2.MaxDiff(to_sample2) < 1e-4);
-                Assert.True(MMath.AbsDiff(evExpected, evActual) < 1e-4);
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "OpenBug")]
-        public void GammaFromShapeAndRateOpTest5()
-        {
-            Gamma sample;
-            Gamma rate;
-            double shape = 1;
-            Gamma q, sampleExpected, sampleActual;
-
-            sample = Gamma.FromShapeAndRate(101, 6.7234079315458819E-154);
-            rate = Gamma.FromShapeAndRate(1, 1);
-            q = GammaFromShapeAndRateOp_Laplace.Q(sample, shape, rate);
-            Console.WriteLine(q);
-            Assert.True(!double.IsNaN(q.Rate));
-            sampleExpected = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-            sampleActual = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(sample, shape, rate, q);
-            Console.WriteLine("sample = {0} should be {1}", sampleActual, sampleExpected);
-            Assert.True(sampleExpected.MaxDiff(sampleActual) < 1e-4);
-
-            sample = Gamma.FromShapeAndRate(1.4616957536444839, 6.2203585601953317E+36);
-            rate = Gamma.FromShapeAndRate(2.5, 0.99222007168007165);
-            sampleExpected = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-            q = Gamma.FromShapeAndRate(3.5, 0.99222007168007154);
-            sampleActual = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(sample, shape, rate, q);
-            Console.WriteLine("sample = {0} should be {1}", sampleActual, sampleExpected);
-            Assert.True(sampleExpected.MaxDiff(sampleActual) < 1e-4);
-
-            sample = Gamma.FromShapeAndRate(1.9692446124520258, 1.0717828357423075E+77);
-            rate = Gamma.FromShapeAndRate(101.0, 2.1709591889324445E-80);
-            sampleExpected = GammaFromShapeAndRateOp_Slow.SampleAverageConditional(sample, shape, rate);
-            q = GammaFromShapeAndRateOp_Laplace.Q(sample, shape, rate);
-            sampleActual = GammaFromShapeAndRateOp_Laplace.SampleAverageConditional(sample, shape, rate, q);
-            Console.WriteLine("sample = {0} should be {1}", sampleActual, sampleExpected);
-            Assert.True(sampleExpected.MaxDiff(sampleActual) < 1e-4);
-
-            Assert.Equal(0.0,
-                            GammaFromShapeAndRateOp_Laplace.LogEvidenceRatio(Gamma.Uniform(), 4.0, Gamma.PointMass(0.01), Gamma.FromShapeAndRate(4, 0.01),
-                                                                             Gamma.PointMass(0.01)));
-        }
-
-        [Fact]
-        public void MatrixTimesVectorOpTest()
-        {
-            Matrix[] ms = new Matrix[3];
-            ms[0] = new Matrix(new double[,] { { 0.036, -0.036, 0 }, { 0.036, 0, -0.036 } });
-            ms[1] = new Matrix(new double[,] { { -0.036, 0.036, 0 }, { 0, 0.036, -0.036 } });
-            ms[2] = new Matrix(new double[,] { { -0.036, 0, 0.036 }, { 0, -0.036, 0.036 } });
-            VectorGaussian product = new VectorGaussian(Vector.FromArray(new double[] { 1, 1 }), PositiveDefiniteMatrix.IdentityScaledBy(2, 0.01));
-            Matrix A = ms[0];
-            VectorGaussian result = MatrixVectorProductOp.BAverageConditional(product, A, new VectorGaussian(3));
-            Assert.False(result.Precision.IsPositiveDefinite());
-
-            Vector BMean = Vector.FromArray(1, 2, 3);
-            PositiveDefiniteMatrix BVariance = new PositiveDefiniteMatrix(new double[,]
-            {
-                { 3,2,1 },
-                { 2,3,2 },
-                { 1,2,3 }
-            });
-            GaussianArray2D to_A = new GaussianArray2D(A.Rows, A.Cols);
-            MatrixVectorProductOp.AAverageConditional(product, GaussianArray2D.PointMass(A.ToArray()), BMean, BVariance, to_A);
-            double[,] dExpected = new double[,]
-            {
-                { 542.014350893474, -34.0730521080406, -261.179402050882 },
-                { 449.735198014871, -31.4263853641523, -215.272881589687 },
-            };
-            for (int i = 0; i < to_A.GetLength(0); i++)
-            {
-                for (int j = 0; j < to_A.GetLength(1); j++)
-                {
-                    Gaussian dist = to_A[i, j];
-                    double dlogp, ddlogp;
-                    dist.GetDerivatives(A[i, j], out dlogp, out ddlogp);
-                    Assert.True(System.Math.Abs(dExpected[i, j] - dlogp) < 1e-8);
-                }
-            }
-        }
-
-        [Fact]
-        public void SumOpTest()
-        {
-            Gaussian sum_F = new Gaussian(0, 1);
-            Gaussian sum_B = Gaussian.FromMeanAndPrecision(0, 1e-310);
-            GaussianArray array = new GaussianArray(2, i => new Gaussian(0, 1));
-            var result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
-            Assert.True(!double.IsNaN(result[0].MeanTimesPrecision));
-
-            sum_B = new Gaussian(0, 1);
-            array[0] = Gaussian.FromMeanAndPrecision(0, 0);
-            sum_F = FastSumOp.SumAverageConditional(array);
-            result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
-            Assert.True(result[0].MaxDiff(new Gaussian(0, 2)) < 1e-10);
-            Assert.True(result[1].IsUniform());
-            for (int i = 0; i < 1030; i++)
-            {
-                array[0] = Gaussian.FromMeanAndPrecision(0, System.Math.Pow(2, -i));
-                sum_F = FastSumOp.SumAverageConditional(array);
-                result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
-                Assert.True(result[0].MaxDiff(new Gaussian(0, 2)) < 1e-10);
-            }
-        }
-
-        //[Fact]
-        //[Trait("Category", "ModifiesGlobals")]
-        internal void ExpOp_CompareToSampling()
-        {
-            int sampleCount = 1_000_000;
-            for (int i = 1; i < 10; i++)
-            {
-                double ve = System.Math.Pow(10, -i);
-                Gamma exp = Gamma.FromMeanAndVariance(1, ve);
-                Gaussian d = Gaussian.FromMeanAndVariance(0, 1);
-                Gaussian to_d = ExpOp.DAverageConditional(exp, d, Gaussian.Uniform());
-                var importanceSampler = new ImportanceSampler(sampleCount, d.Sample, x => System.Math.Exp(exp.GetLogProb(System.Math.Exp(x))));
-                double mean = importanceSampler.GetExpectation(x => x);
-                double Ex2 = importanceSampler.GetExpectation(x => x * x);
-                Gaussian to_d_is = new Gaussian(mean, Ex2 - mean * mean) / d;
-                var exp2 = Gamma.FromShapeAndRate(exp.Shape - 1, exp.Rate);
-                importanceSampler = new ImportanceSampler(sampleCount, exp2.Sample, x => System.Math.Exp(d.GetLogProb(System.Math.Log(x))));
-                mean = importanceSampler.GetExpectation(x => System.Math.Log(x));
-                Ex2 = importanceSampler.GetExpectation(x => System.Math.Log(x) * System.Math.Log(x));
-                Gaussian to_d_is2 = new Gaussian(mean, Ex2 - mean * mean) / d;
-                Trace.WriteLine($"{ve}\t{to_d.GetMean()}\t{to_d_is.GetMean()}\t{to_d_is2.GetMean()}");
-
-                Gamma to_exp = ExpOp.ExpAverageConditional(exp, d, Gaussian.Uniform());
-
-                importanceSampler = new ImportanceSampler(sampleCount, d.Sample, x => System.Math.Exp(exp.GetLogProb(System.Math.Exp(x))));
-                double Ey = importanceSampler.GetExpectation(x => System.Math.Exp(x));
-                double Elogy = importanceSampler.GetExpectation(x => x);
-                var to_exp_is1 = Gamma.FromMeanAndMeanLog(Ey, Elogy) / exp;
-                importanceSampler = new ImportanceSampler(sampleCount, exp2.Sample, x => System.Math.Exp(d.GetLogProb(System.Math.Log(x))));
-                Ey = importanceSampler.GetExpectation(x => x);
-                Elogy = importanceSampler.GetExpectation(x => System.Math.Log(x));
-                var to_exp_is2 = Gamma.FromMeanAndMeanLog(Ey, Elogy) / exp;
-                double is1_mean = to_exp_is1.IsProper() ? to_exp_is1.GetMean() : double.NaN;
-                double is2_mean = to_exp_is2.IsProper() ? to_exp_is2.GetMean() : double.NaN;
-                Trace.WriteLine($"to_exp: Quad {to_exp.GetMean()} IS1: {is1_mean} IS2: {is2_mean}");
-            }
-        }
-
-        [Fact]
-        public void ExpOpGammaPowerTest()
-        {
-            Assert.True(!double.IsNaN(ExpOp.ExpAverageConditional(GammaPower.Uniform(-1), Gaussian.FromNatural(0.046634157098979417, 0.00078302234897204242), Gaussian.Uniform()).Rate));
-            Assert.True(!double.IsNaN(ExpOp.ExpAverageConditional(GammaPower.Uniform(-1), Gaussian.FromNatural(0.36153121930654075, 0.0005524890062312658), Gaussian.Uniform()).Rate));
-            Assert.True(ExpOp.DAverageConditional(GammaPower.PointMass(0, -1), new Gaussian(0, 1), Gaussian.Uniform()).Point < double.MinValue);
-            ExpOp.ExpAverageConditional(GammaPower.FromShapeAndRate(-1, 283.673, -1), Gaussian.FromNatural(0.004859823703146038, 6.6322755562737905E-06), Gaussian.FromNatural(0.00075506803981220758, 8.24487022054953E-07));
-            GammaPower exp = GammaPower.FromShapeAndRate(0, 0, -1);
-            Gaussian[] ds = new[]
-            {
-                Gaussian.FromNatural(-1.6171314269768655E+308, 4.8976001759138024),
-                Gaussian.FromNatural(-0.037020622891705768, 0.00034989765084474117),
-                Gaussian.PointMass(double.NegativeInfinity),
-            };
-            foreach (var d in ds)
-            {
-                Gaussian to_d = ExpOp.DAverageConditional(exp, d, Gaussian.Uniform());
-                Gaussian to_d_slow = ExpOp_Slow.DAverageConditional(exp, d);
-                //Trace.WriteLine($"{to_d}");
-                //Trace.WriteLine($"{to_d_slow}");
-                Assert.True(to_d_slow.MaxDiff(to_d) < 1e-10);
-                to_d = Gaussian.FromNatural(1, 0);
-                GammaPower to_exp = ExpOp.ExpAverageConditional(exp, d, to_d);
-                //Trace.WriteLine($"{to_exp}");
-            }
-            ExpOp.ExpAverageConditional(GammaPower.FromShapeAndRate(-1, 883.22399999999993, -1), Gaussian.FromNatural(0.0072160312702854888, 8.1788482512051846E-06), Gaussian.FromNatural(0.00057861649495666474, 5.6316164560235272E-07));
-        }
-
-        [Fact]
-        [Trait("Category", "OpenBug")]
-        public void ExpOpTest()
-        {
-            Assert.True(ExpOp.ExpAverageConditional(Gamma.FromShapeAndRate(3.302758272196654, 0.00060601537137241492), Gaussian.FromNatural(55.350150233321628, 6.3510247863590683), Gaussian.FromNatural(27.960892513144643, 3.4099170930572216)).Rate > 0);
-            Gamma exp = new Gamma(1, 1);
-            Gaussian[] ds = new[]
-            {
-                Gaussian.FromNatural(-1.6171314269768655E+308, 4.8976001759138024),
-                Gaussian.PointMass(double.NegativeInfinity),
-            };
-            foreach (var d in ds)
-            {
-                Gamma to_exp = ExpOp.ExpAverageConditional(exp, d, Gaussian.Uniform());
-                Gaussian to_d = ExpOp.DAverageConditional(exp, d, Gaussian.Uniform());
-                Gaussian to_d_slow = ExpOp_Slow.DAverageConditional(exp, d);
-                Trace.WriteLine($"{to_d}");
-                Trace.WriteLine($"{to_d_slow}");
-                Assert.True(to_d_slow.MaxDiff(to_d) < 1e-10);
-            }
-        }
-
-        // This test fails because of cancellation in logMeanMinusMd and Gamma.SetToRatio in ExpOp.ExpAverageConditional
-        [Fact]
-        [Trait("Category", "ModifiesGlobals")]
-        [Trait("Category", "OpenBug")]
-        public void ExpOp_PointExp()
-        {
-            using (TestUtils.TemporarilyChangeQuadratureNodeCount(21))
-            {
-                double vd = 1e-4;
-                vd = 1e-3;
-                Gaussian d = new Gaussian(0, vd);
-                double ve = 2e-3;
-                //ve = 1;
-                Gaussian uniform = Gaussian.Uniform();
-                Gamma expPoint = Gamma.PointMass(2);
-                Gamma to_exp_point = ExpOp.ExpAverageConditional(expPoint, d, uniform);
-                Gaussian to_d_point = ExpOp.DAverageConditional(expPoint, d, uniform);
-                double to_exp_oldError = double.PositiveInfinity;
-                double to_d_oldError = double.PositiveInfinity;
-                for (int i = 5; i < 100; i++)
-                {
-                    ve = System.Math.Pow(10, -i);
-                    Gamma exp = Gamma.FromMeanAndVariance(2, ve);
-                    Gamma to_exp = ExpOp.ExpAverageConditional(exp, d, uniform);
-                    Gaussian to_d = ExpOp.DAverageConditional(exp, d, uniform);
-                    double to_exp_error = to_exp.MaxDiff(to_exp_point);
-                    double to_d_error = System.Math.Abs(to_d.GetMean() - to_d_point.GetMean());
-                    Trace.WriteLine($"ve={ve}: to_exp={to_exp} error={to_exp_error} to_d={to_d} error={to_d_error}");
-                    Assert.True(to_exp_error <= to_exp_oldError);
-                    to_exp_oldError = to_exp_error;
-                    Assert.True(to_d_error <= to_d_oldError);
-                    to_d_oldError = to_d_error;
-                }
-                Trace.WriteLine(ExpOp.DAverageConditional(Gamma.FromMeanAndVariance(1, ve), d, uniform));
-                using (TestUtils.TemporarilyChangeQuadratureShift(true))
-                {
-                    Trace.WriteLine(ExpOp.DAverageConditional(Gamma.FromMeanAndVariance(1, ve), d, uniform));
-                }
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "OpenBug")]
-        public void ExpOpGammaPower_PointExp()
-        {
-            double power = -1;
-            double vd = 1e-4;
-            vd = 1e-3;
-            Gaussian d = new Gaussian(0, vd);
-            Gaussian uniform = Gaussian.Uniform();
-            GammaPower expPoint = GammaPower.PointMass(2, power);
-            GammaPower to_exp_point = ExpOp.ExpAverageConditional(expPoint, d, uniform);
-            Gaussian to_d_point = ExpOp.DAverageConditional(expPoint, d, uniform);
-            double to_exp_oldError = double.PositiveInfinity;
-            double to_d_oldError = double.PositiveInfinity;
-            for (int i = 0; i < 100; i++)
-            {
-                double ve = System.Math.Pow(10, -i);
-                GammaPower exp = GammaPower.FromMeanAndVariance(2, ve, power);
-                GammaPower to_exp = ExpOp.ExpAverageConditional(exp, d, uniform);
-                Gaussian to_d = ExpOp.DAverageConditional(exp, d, uniform);
-                double to_exp_error = to_exp.MaxDiff(to_exp_point);
-                double to_d_error = System.Math.Abs(to_d.GetMean() - to_d_point.GetMean());
-                Trace.WriteLine($"ve={ve}: to_exp={to_exp} error={to_exp_error} to_d={to_d} error={to_d_error}");
-                Assert.True(to_exp_error <= to_exp_oldError);
-                to_exp_oldError = to_exp_error;
-                Assert.True(to_d_error <= to_d_oldError);
-                to_d_oldError = to_d_error;
-            }
-        }
-
-        [Fact]
-        public void LogisticOpTest()
-        {
-            for (int trial = 0; trial < 2; trial++)
-            {
-                double xMean = (trial == 0) ? -1 : 1;
-                Gaussian x = Gaussian.FromMeanAndVariance(xMean, 0);
-                Gaussian result2 = BernoulliFromLogOddsOp.LogOddsAverageConditional(true, x);
-                Console.WriteLine("{0}: {1}", x, result2);
-                for (int i = 8; i < 30; i++)
-                {
-                    double v = System.Math.Pow(0.1, i);
-                    x = Gaussian.FromMeanAndVariance(xMean, v);
-                    Gaussian result = BernoulliFromLogOddsOp.LogOddsAverageConditional(true, x);
-                    Console.WriteLine("{0}: {1} maxDiff={2}", x, result, result2.MaxDiff(result));
-                    Assert.True(result2.MaxDiff(result) < 1e-6);
-                }
-            }
-
-            for (int i = 0; i < 10; i++)
-            {
-                Gaussian falseMsg = LogisticOp.FalseMsg(new Beta(0.2, 1.8), new Gaussian(0, 97.0 * (i + 1)), new Gaussian());
-                Console.WriteLine(falseMsg);
-            }
-            Gaussian toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(-4662, 1314));
-            Console.WriteLine(toLogOdds);
-            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(2249, 2.5));
-            Console.WriteLine(toLogOdds);
-            Gaussian logOdds = new Gaussian(100, 100);
-            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, logOdds);
-            Console.WriteLine(toLogOdds * logOdds);
-            // test m =approx 1.5*v for increasing v
-            for (int i = 0; i < 10; i++)
-            {
-                double v = System.Math.Pow(2, i + 5);
-                Gaussian g = new Gaussian(37 + 1.5 * v, v);
-                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, g);
-                Console.WriteLine("{0}: {1}", g, toLogOdds);
-                Gaussian actualPost = toLogOdds * g;
-                Gaussian expectedPost = new Gaussian(0, 1);
-                if (i == 0)
-                {
-                    // for Gaussian(85,32)
-                    expectedPost = new Gaussian(53, 32);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-                else if (i == 1)
-                {
-                    // for Gaussian(133,64)
-                    expectedPost = new Gaussian(69, 64);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-                else if (i == 4)
-                {
-                    // for Gaussian(805,512) the posterior should be (293, 511.81)
-                    expectedPost = new Gaussian(293, 511.81);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-                else if (i == 5)
-                {
-                    // for Gaussian(1573,1024)
-                    expectedPost = new Gaussian(549, 1023.9);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-            }
-            // test m =approx v for increasing v
-            for (int i = 0; i < 10; i++)
-            {
-                double v = System.Math.Pow(2, i + 5);
-                Gaussian g = new Gaussian(v - 1, v);
-                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, g);
-                Console.WriteLine("{0}: {1}", g, toLogOdds);
-                Gaussian actualPost = toLogOdds * g;
-                Gaussian expectedPost = new Gaussian(0, 1);
-                if (i == 0)
-                {
-                    // for Gaussian(31,32)
-                    expectedPost = new Gaussian(3.8979, 12.4725);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-                else if (i == 6)
-                {
-                    // for Gaussian(2047,2048)
-                    expectedPost = new Gaussian(35.7156, 736.2395);
-                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
-                }
-            }
-            for (int i = 0; i < 10; i++)
-            {
-                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(54.65 / 10 * (i + 1), 8.964));
-                Console.WriteLine(toLogOdds);
-            }
-            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(9900, 10000) ^ 0.1);
-            Console.WriteLine(toLogOdds);
-            Gaussian falseMsg2 = LogisticOp.FalseMsg(new Beta(0.9, 0.1), Gaussian.FromNatural(-10.097766458353044, 0.000011644704327819733),
-              Gaussian.FromNatural(-0.0010832099815010626, 0.000010092906656322242));
-            Console.WriteLine(falseMsg2);
-        }
-
-        [Fact]
-        public void BernoulliFromLogOddsTest()
-        {
-            double tolerance = 1e-8;
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(0, 2)), System.Math.Log(0.5)) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(2, 0)), MMath.LogisticLn(2)) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(1, 1)), System.Math.Log(1 - 0.5 * System.Math.Exp(-0.5))) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(10, 10)), System.Math.Log(1 - 0.5 * System.Math.Exp(-5))) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-1, 1)), System.Math.Log(0.303265329856008)) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-5, 3)), System.Math.Log(0.023962216989475)) < tolerance);
-            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-5, 10)), System.Math.Log(0.084396512538678)) < tolerance);
-        }
-
-        [Fact]
-        public void MaxTest()
-        {
-            Gaussian actual, expected;
-            actual = MaxGaussianOp.MaxAverageConditional(Gaussian.FromNatural(6053.7946407740192, 2593.4559834344436), Gaussian.FromNatural(-1.57090676324773, 1.3751262174888785), Gaussian.FromNatural(214384.78500926663, 96523.508973471908));
-            Assert.False(actual.IsProper());
-            actual = MaxGaussianOp.MaxAverageConditional(Gaussian.FromNatural(146.31976467723146, 979.371757950659), Gaussian.FromNatural(0.075442729439046508, 0.086399540048904114), Gaussian.PointMass(0));
-            Assert.False(actual.IsProper());
-
-            actual = MaxGaussianOp.MaxAverageConditional(Gaussian.Uniform(), Gaussian.PointMass(0), new Gaussian(-7.357e+09, 9.75));
-            expected = Gaussian.PointMass(0);
-            Assert.True(expected.MaxDiff(actual) < 1e-4);
-
-            Gaussian max;
-            max = new Gaussian(4, 5);
-            actual = MaxGaussianOp.MaxAverageConditional(max, new Gaussian(0, 1), new Gaussian(2, 3));
-            actual *= max;
-            expected = new Gaussian(2.720481395499785, 1.781481142817509);
-            //expected = MaxPosterior(max, new Gaussian(0, 1), new Gaussian(2, 3));
-            Assert.True(expected.MaxDiff(actual) < 1e-4);
-
-            max = new Gaussian();
-            max.MeanTimesPrecision = 0.2;
-            max.Precision = 1e-10;
-            actual = MaxGaussianOp.MaxAverageConditional(max, new Gaussian(0, 1), new Gaussian(0, 1));
-            actual *= max;
-            expected = new Gaussian(0.702106815765215, 0.697676918460236);
-            Assert.True(expected.MaxDiff(actual) < 1e-4);
-        }
-
-        [Fact]
-        public void MaxTest2()
-        {
-            foreach (double max in new[] { 0.0, 2.0 })
-            {
-                double oldm = double.NaN;
-                double oldv = double.NaN;
-                for (int i = 0; i < 300; i++)
-                {
-                    Gaussian a = new Gaussian(System.Math.Pow(10, i), 177);
-                    Gaussian to_a = MaxGaussianOp.AAverageConditional(max, a, 0);
-                    Gaussian to_b = MaxGaussianOp.BAverageConditional(max, 0, a);
-                    Assert.Equal(to_a, to_b);
-                    if (max == 0)
-                    {
-                        Gaussian to_a2 = IsPositiveOp.XAverageConditional(false, a);
-                        double error = System.Math.Max(MMath.AbsDiff(to_a.MeanTimesPrecision, to_a2.MeanTimesPrecision, double.Epsilon),
-                            MMath.AbsDiff(to_a.Precision, to_a2.Precision, double.Epsilon));
-                        //Trace.WriteLine($"{a} {to_a} {to_a2} {error}");
-                        Assert.True(error < 1e-12);
-                    }
-                    //else Trace.WriteLine($"{a} {to_a}");
-                    double m, v;
-                    to_a.GetMeanAndVariance(out m, out v);
-                    if (!double.IsNaN(oldm))
-                    {
-                        Assert.True(v <= oldv);
-                        double olddiff = System.Math.Abs(max - oldm);
-                        double diff = System.Math.Abs(max - m);
-                        Assert.True(diff <= olddiff);
-                    }
-                    oldm = m;
-                    oldv = v;
-                }
-            }
-        }
-
-        [Fact]
-        public void Max_MaxPointMassTest()
-        {
-            Gaussian a = new Gaussian(1, 2);
-            Gaussian b = new Gaussian(4, 5);
-            Max_MaxPointMass(a, b);
-            Max_MaxPointMass(a, Gaussian.PointMass(2));
-            Max_MaxPointMass(a, Gaussian.PointMass(3));
-            Max_MaxPointMass(Gaussian.PointMass(2), b);
-            Max_MaxPointMass(Gaussian.PointMass(3), b);
-            Max_MaxPointMass(Gaussian.PointMass(3), Gaussian.PointMass(2));
-            Max_MaxPointMass(Gaussian.PointMass(2), Gaussian.PointMass(3));
-        }
-
-        private void Max_MaxPointMass(Gaussian a, Gaussian b)
-        {
-            double point = 3;
-            Gaussian toPoint = MaxGaussianOp.MaxAverageConditional(Gaussian.PointMass(point), a, b);
-            //Console.WriteLine($"{point} {toPoint} {toPoint.MeanTimesPrecision} {toPoint.Precision}");
-            double oldDiff = double.PositiveInfinity;
-            for (int i = 5; i < 100; i++)
-            {
-                Gaussian max = Gaussian.FromMeanAndPrecision(point, System.Math.Pow(10, i));
-                Gaussian to_max = MaxGaussianOp.MaxAverageConditional(max, a, b);
-                double diff = toPoint.MaxDiff(to_max);
-                //Console.WriteLine($"{max} {to_max} {to_max.MeanTimesPrecision} {to_max.Precision} {diff}");
-                if (diff < 1e-14) diff = 0;
-                Assert.True(diff <= oldDiff);
-                oldDiff = diff;
-            }
-        }
-
-        [Fact]
-        public void Max_APointMassTest()
-        {
-            //MaxGaussianOp.ForceProper = false;
-            Gaussian max = new Gaussian(4, 5);
-            Gaussian b = new Gaussian(1, 2);
-            Max_APointMass(new Gaussian(4, 1e-0), Gaussian.PointMass(0));
-            Max_APointMass(Gaussian.PointMass(11), Gaussian.PointMass(0));
-            Max_APointMass(max, b);
-            Max_APointMass(max, Gaussian.PointMass(2));
-            Max_APointMass(max, Gaussian.PointMass(3));
-            Max_APointMass(max, Gaussian.PointMass(4));
-            Max_APointMass(Gaussian.PointMass(3), b);
-            Max_APointMass(Gaussian.PointMass(4), b);
-            Max_APointMass(Gaussian.PointMass(3), Gaussian.PointMass(3));
-        }
-
-        private void Max_APointMass(Gaussian max, Gaussian b)
-        {
-            double point = 3;
-            Gaussian toPoint = MaxGaussianOp.AAverageConditional(max, Gaussian.PointMass(point), b);
-            //Console.WriteLine($"{point} {toPoint} {toPoint.MeanTimesPrecision:g17} {toPoint.Precision:g17}");
-            if (max.IsPointMass && b.IsPointMass)
-            {
-                Gaussian toUniform = MaxGaussianOp.AAverageConditional(max, Gaussian.Uniform(), b);
-                if (max.Point > b.Point)
-                {
-                    Assert.Equal(toUniform, max);
-                }
-                else
-                {
-                    Assert.Equal(toUniform, Gaussian.Uniform());
-                }
-            }
-            double oldDiff = double.PositiveInfinity;
-            for (int i = 3; i < 100; i++)
-            {
-                Gaussian a = Gaussian.FromMeanAndPrecision(point, System.Math.Pow(10, i));
-                Gaussian to_a = MaxGaussianOp.AAverageConditional(max, a, b);
-                double diff = toPoint.MaxDiff(to_a);
-                //Console.WriteLine($"{a} {to_a} {to_a.MeanTimesPrecision:g17} {to_a.Precision:g17} {diff:g17}");
-                if (diff < 1e-14) diff = 0;
-                Assert.True(diff <= oldDiff);
-                oldDiff = diff;
-            }
-        }
-
-        public static Gaussian MaxAPosterior(Gaussian max, Gaussian a, Gaussian b)
-        {
-            int n = 10000000;
-            GaussianEstimator est = new GaussianEstimator();
-            for (int i = 0; i < n; i++)
-            {
-                double aSample = a.Sample();
-                double bSample = b.Sample();
-                double logProb = max.GetLogProb(System.Math.Max(aSample, bSample));
-                double weight = System.Math.Exp(logProb);
-                est.Add(aSample, weight);
-            }
-            return est.GetDistribution(new Gaussian());
-        }
-
-        public static Gaussian MaxPosterior(Gaussian max, Gaussian a, Gaussian b)
-        {
-            int n = 10000000;
-            GaussianEstimator est = new GaussianEstimator();
-            for (int i = 0; i < n; i++)
-            {
-                double aSample = a.Sample();
-                double bSample = b.Sample();
-                double maxSample = System.Math.Max(aSample, bSample);
-                double logProb = max.GetLogProb(maxSample);
-                double weight = System.Math.Exp(logProb);
-                est.Add(maxSample, weight);
-            }
-            return est.GetDistribution(new Gaussian());
-        }
-
-        [Fact]
-        public void OrTest()
-        {
-            Assert.Equal(1.0, BooleanOrOp.AAverageConditional(true, false).GetProbTrue());
-            Assert.Equal(1.0, BooleanOrOp.OrAverageConditional(Bernoulli.PointMass(true), false).GetProbTrue());
-
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), Bernoulli.PointMass(false)).LogOdds);
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, Bernoulli.PointMass(false)).LogOdds);
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), false).LogOdds);
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, false).LogOdds);
-
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), Bernoulli.PointMass(true)).GetProbTrue());
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, Bernoulli.PointMass(true)).GetProbTrue());
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), true).GetProbTrue());
-            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, true).GetProbTrue());
-
-            Assert.Throws<AllZeroException>(() =>
-            {
-                BooleanAndOp.AAverageConditional(true, false);
-            });
-            Assert.Throws<AllZeroException>(() =>
-            {
-                BooleanAndOp.AAverageConditional(Bernoulli.PointMass(true), false);
-            });
-            Assert.Throws<AllZeroException>(() =>
-            {
-                BooleanAndOp.AAverageConditional(true, Bernoulli.PointMass(false));
-            });
-            Assert.Throws<AllZeroException>(() =>
-            {
-                BooleanAndOp.AAverageConditional(Bernoulli.PointMass(true), Bernoulli.PointMass(false));
-            });
-        }
-
-        [Fact]
-        public void BernoulliFromDiscreteTest()
-        {
-            double[] probTrue = { 0.4, 0.7 };
-            Discrete indexDist = new Discrete(0.8, 0.2);
-            Bernoulli sampleDist = new Bernoulli(0.8 * 0.4 + 0.2 * 0.7);
-            Assert.True(sampleDist.MaxDiff(BernoulliFromDiscreteOp.SampleAverageConditional(indexDist, probTrue)) < 1e-4);
-            sampleDist.SetProbTrue(0.2);
-            double p = (0.8 * 0.3 + 0.2 * 0.7) / (0.8 * 0.3 + 0.2 * 0.7 + 0.8 * 0.6 + 0.2 * 0.4);
-            Vector probs = indexDist.GetWorkspace();
-            probs[0] = 1 - p;
-            probs[1] = p;
-            indexDist.SetProbs(probs);
-            Assert.True(indexDist.MaxDiff(BernoulliFromDiscreteOp.IndexAverageConditional(sampleDist, probTrue, Discrete.Uniform(indexDist.Dimension))) < 1e-4);
-        }
-
-        [Fact]
-        public void BernoulliFromBooleanTest()
-        {
-            double[] probTrue = { 0.4, 0.7 };
-            Bernoulli choiceDist = new Bernoulli(0.2);
-            Bernoulli sampleDist = new Bernoulli(0.8 * 0.4 + 0.2 * 0.7);
-            Assert.True(sampleDist.MaxDiff(BernoulliFromBooleanArray.SampleAverageConditional(choiceDist, probTrue)) < 1e-4);
-            sampleDist.SetProbTrue(0.2);
-            choiceDist.SetProbTrue((0.8 * 0.3 + 0.2 * 0.7) / (0.8 * 0.3 + 0.2 * 0.7 + 0.8 * 0.6 + 0.2 * 0.4));
-            Assert.True(choiceDist.MaxDiff(BernoulliFromBooleanArray.ChoiceAverageConditional(sampleDist, probTrue)) < 1e-4);
-        }
-
-        [Fact]
-        [Trait("Category", "ModifiesGlobals")]
-        public void BernoulliFromBetaOpTest()
-        {
-            Assert.True(System.Math.Abs(BernoulliFromBetaOp.LogEvidenceRatio(new Bernoulli(3e-5), Beta.PointMass(1)) - 0) < 1e-10);
-
-            using (TestUtils.TemporarilyAllowBetaImproperSums)
-            {
-                Beta probTrueDist = new Beta(3, 2);
-                Bernoulli sampleDist = new Bernoulli();
-                Assert.True(new Beta(1, 1).MaxDiff(BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist)) < 1e-4);
-                sampleDist = Bernoulli.PointMass(true);
-                Assert.True(new Beta(2, 1).MaxDiff(BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist)) < 1e-4);
-                sampleDist = Bernoulli.PointMass(false);
-                Assert.True(new Beta(1, 2).MaxDiff(BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist)) < 1e-4);
-                sampleDist = new Bernoulli(0.9);
-                Assert.True(new Beta(1.724, 0.9598).MaxDiff(BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist)) < 1e-3);
-                Assert.Throws<ImproperMessageException>(() =>
-                {
-                    BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, new Beta(1, -2));
-                });
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "ModifiesGlobals")]
-        public void BernoulliFromBetaOpTest2()
-        {
-            Bernoulli sampleDist = new Bernoulli(2.0 / 3);
-            using (TestUtils.TemporarilyAllowBetaImproperSums)
-            {
-                for (int i = 10; i <= 10; i++)
-                {
-                    double s = System.Math.Exp(i);
-                    Beta probTrueDist = new Beta(s, 2 * s);
-                    Beta expected = BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist);
-                    if (i == 10)
-                        probTrueDist.Point = probTrueDist.GetMean();
-                    Beta actual = BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist);
-                    Console.WriteLine("{0} {1}", probTrueDist, actual);
-                    Assert.True(expected.MaxDiff(actual) < 1e-4);
-                }
-            }
-            for (int i = 10; i <= 10; i++)
-            {
-                Beta probTrueDist = new Beta(System.Math.Exp(i), 1);
-                Beta expected = BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist);
-                if (i == 10)
-                    probTrueDist.Point = 1;
-                Beta actual = BernoulliFromBetaOp.ProbTrueAverageConditional(sampleDist, probTrueDist);
-                Console.WriteLine("{0} {1}", probTrueDist, actual);
-                Assert.True(expected.MaxDiff(actual) < 1e-4);
-            }
-        }
-
-        [Trait("Category", "ModifiesGlobals")]
-        internal void DiscreteFromDirichletOpTest2()
-        {
-            Discrete sample = new Discrete(1.0 / 3, 2.0 / 3);
-            Vector sampleProbs = sample.GetProbs();
-            Dirichlet result = Dirichlet.Uniform(2);
-            using (TestUtils.TemporarilyAllowDirichletImproperSums)
-            {
-                for (int i = 1; i <= 10; i++)
-                {
-                    double s = System.Math.Exp(i);
-                    Dirichlet probsDist = new Dirichlet(s, 2 * s);
-                    if (i == 10)
-                        probsDist.Point = probsDist.GetMean();
-                    Console.WriteLine("{0} {1}", probsDist, DiscreteFromDirichletOp.ProbsAverageConditional(sample, probsDist, result));
-                    Dirichlet post = probsDist * result;
-                    Vector postMean = post.GetMean();
-                    Vector priorMean = probsDist.GetMean();
-                    Vector postMeanSquare = post.GetMeanSquare();
-                    DenseVector delta2 = DenseVector.Zero(postMean.Count);
-                    for (int j = 0; j < delta2.Count; j++)
-                    {
-                        delta2[j] = postMeanSquare[j] - (probsDist.PseudoCount[j] + 1) / (probsDist.TotalCount + 1) * postMean[j];
-                    }
-                    delta2.Scale(probsDist.TotalCount + 1);
-                    Vector delta1 = (postMean - priorMean) * probsDist.TotalCount;
-                    DenseVector delta1Approx = DenseVector.Zero(postMean.Count);
-                    //delta.SetToFunction(sampleProbs, priorMean, (sampleProb, prob) => (
-                    double Z = sampleProbs.Inner(priorMean);
-                    delta1Approx[0] = priorMean[0] * priorMean[1] * (sampleProbs[0] - sampleProbs[1]) / Z;
-                    DenseVector delta2Approx = DenseVector.Zero(postMean.Count);
-                    delta2Approx[0] = priorMean[0] * priorMean[0] * priorMean[1] * (sampleProbs[0] - sampleProbs[1]) / Z;
-                    //Console.WriteLine("delta = {0} {1} {2} {3}", delta1, delta1Approx, delta2, delta2Approx);
-                }
-            }
-            for (int i = 1; i <= 10; i++)
-            {
-                double s = System.Math.Exp(i);
-                Dirichlet probsDist = new Dirichlet(s * 4, s * 5);
-                if (i == 10)
-                    probsDist.Point = probsDist.GetMean();
-                Console.WriteLine("{0} {1}", probsDist, DiscreteFromDirichletOp.ProbsAverageConditional(sample, probsDist, result));
-            }
-        }
-
-        [Trait("Category", "ModifiesGlobals")]
-        internal void DiscreteFromDirichletOpTest3()
-        {
-            Discrete sample = new Discrete(3.0 / 6, 2.0 / 6, 1.0 / 6);
-            Vector sampleProbs = sample.GetProbs();
-            Dirichlet result = Dirichlet.Uniform(3);
-            using (TestUtils.TemporarilyAllowDirichletImproperSums)
-            {
-                for (int i = 1; i <= 10; i++)
-                {
-                    double s = System.Math.Exp(i);
-                    Dirichlet probsDist = new Dirichlet(s * 2, s * 1, s * 3);
-                    if (i == 10)
-                        probsDist.Point = probsDist.GetMean();
-                    Console.WriteLine("{0} {1}", probsDist, DiscreteFromDirichletOp.ProbsAverageConditional(sample, probsDist, result));
-                    Dirichlet post = probsDist * result;
-                    Vector postMean = post.GetMean();
-                    Vector postMeanSquare = post.GetMeanSquare();
-                    Vector priorMean = probsDist.GetMean();
-                    Vector delta1 = (postMean - priorMean) * probsDist.TotalCount;
-                    DenseVector delta2 = DenseVector.Zero(postMean.Count);
-                    for (int j = 0; j < delta2.Count; j++)
-                    {
-                        delta2[j] = postMeanSquare[j] - (probsDist.PseudoCount[j] + 1) / (probsDist.TotalCount + 1) * postMean[j];
-                    }
-                    delta2.Scale(probsDist.TotalCount + 1);
-                    double Z = sampleProbs.Inner(priorMean);
-                    DenseVector delta1Approx = DenseVector.Zero(postMean.Count);
-                    //delta.SetToFunction(sampleProbs, priorMean, (sampleProb, prob) => (
-                    //delta1Approx[0] = priorMean[0] * priorMean[2] * (sampleProbs[0] - sampleProbs[2])/Z;
-                    delta1Approx[0] = probsDist.PseudoCount[1] * postMean[0] - probsDist.PseudoCount[0] * postMean[1] + priorMean[0] * priorMean[2] * (sampleProbs[0] - sampleProbs[2]) / Z;
-                    delta1Approx[0] = (probsDist.PseudoCount[1] + probsDist.PseudoCount[2]) / probsDist.PseudoCount[2] * priorMean[0] * priorMean[2] * (sampleProbs[0] - sampleProbs[2]) / Z
-                        - probsDist.PseudoCount[0] / probsDist.PseudoCount[2] * priorMean[1] * priorMean[2] * (sampleProbs[1] - sampleProbs[2]) / Z;
-                    delta1Approx[1] = probsDist.PseudoCount[0] * postMean[1] - probsDist.PseudoCount[1] * postMean[0] + priorMean[1] * priorMean[2] * (sampleProbs[1] - sampleProbs[2]) / Z;
-                    delta1Approx[2] = -priorMean[0] * priorMean[2] * (sampleProbs[0] - sampleProbs[2]) / Z - priorMean[1] * priorMean[2] * (sampleProbs[1] - sampleProbs[2]) / Z;
-                    DenseVector delta2Approx = DenseVector.Zero(postMean.Count);
-                    delta2Approx[0] = priorMean[0] * priorMean[0] * priorMean[1] * (sampleProbs[0] - sampleProbs[1]) / Z;
-                    //Console.WriteLine("delta = {0} {1} {2} {3}", delta1, delta1Approx, delta2, delta2Approx);
-                }
-            }
-            for (int i = 1; i <= 10; i++)
-            {
-                double s = System.Math.Exp(i);
-                Dirichlet probsDist = new Dirichlet(s * 4, s * 5, s * 6);
-                if (i == 10)
-                    probsDist.Point = probsDist.GetMean();
-                Console.WriteLine("{0} {1}", probsDist, DiscreteFromDirichletOp.ProbsAverageConditional(sample, probsDist, result));
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "ModifiesGlobals")]
-        public void DiscreteFromDirichletOpMomentMatchTest()
-        {
-            using (TestUtils.TemporarilyAllowDirichletImproperSums)
-            {
-                Dirichlet probsDist = new Dirichlet(1, 2, 3, 4);
-                Discrete sampleDist = Discrete.Uniform(4);
-                Assert.True(Dirichlet.Uniform(4).MaxDiff(DiscreteFromDirichletOp.ProbsAverageConditional(sampleDist, probsDist, Dirichlet.Uniform(4))) < 1e-4);
-                sampleDist = Discrete.PointMass(1, 4);
-                Assert.True(new Dirichlet(1, 2, 1, 1).MaxDiff(DiscreteFromDirichletOp.ProbsAverageConditional(sampleDist, probsDist, Dirichlet.Uniform(4))) < 1e-4);
-                sampleDist = new Discrete(0, 1, 1, 0);
-                Assert.True(new Dirichlet(0.9364, 1.247, 1.371, 0.7456).MaxDiff(DiscreteFromDirichletOp.ProbsAverageConditional(sampleDist, probsDist, Dirichlet.Uniform(4))) <
-                              1e-3);
-            }
-        }
-
-        [Fact]
-        public void DiscreteFromDirichletOpTest()
-        {
-            Dirichlet probs4 = new Dirichlet(1.0, 2, 3, 4);
-            Discrete sample4 = new Discrete(0.4, 0.6, 0, 0);
-            Dirichlet result4 = DiscreteFromDirichletOp.ProbsAverageConditional(sample4, probs4, Dirichlet.Uniform(4));
-
-            Dirichlet probs3 = new Dirichlet(1.0, 2, 7);
-            Discrete sample3 = new Discrete(0.4, 0.6, 0);
-            Dirichlet result3 = DiscreteFromDirichletOp.ProbsAverageConditional(sample3, probs3, Dirichlet.Uniform(3));
-            for (int i = 0; i < 3; i++)
-            {
-                Assert.True(MMath.AbsDiff(result4.PseudoCount[i], result3.PseudoCount[i], 1e-6) < 1e-10);
-            }
-
-            Dirichlet probs2 = new Dirichlet(1.0, 2);
-            Discrete sample2 = new Discrete(0.4, 0.6);
-            Dirichlet result2 = DiscreteFromDirichletOp.ProbsAverageConditional(sample2, probs2, Dirichlet.Uniform(2));
-
-            Beta beta = new Beta(1.0, 2);
-            Bernoulli bernoulli = new Bernoulli(0.4);
-            Beta result1 = BernoulliFromBetaOp.ProbTrueAverageConditional(bernoulli, beta);
-            Assert.Equal(result2.PseudoCount[0], result1.TrueCount, 1e-10);
-            Assert.Equal(result2.PseudoCount[1], result1.FalseCount, 1e-10);
-
-            // test handling of small alphas
-            Discrete sample = DiscreteFromDirichletOp.SampleAverageLogarithm(Dirichlet.Symmetric(2, 1e-8), Discrete.Uniform(2));
-            Assert.True(sample.IsUniform());
-        }
-
-        // test the limit of an incoming message with small precision
-        [Fact]
-        public void GaussianIsPositiveTest2()
-        {
-            for (int trial = 0; trial < 2; trial++)
-            {
-                bool isPositive = (trial == 0);
-                Gaussian x = Gaussian.FromNatural(isPositive ? -2 : 2, 0);
-                Gaussian result = IsPositiveOp.XAverageConditional(isPositive, x);
-                for (int i = 10; i < 20; i++)
-                {
-                    x.Precision = System.Math.Pow(0.1, i);
-                    Gaussian result2 = IsPositiveOp.XAverageConditional(isPositive, x);
-                    //Console.WriteLine($"{x}: {result2} maxDiff={result.MaxDiff(result2)}");
-                    Assert.True(result.MaxDiff(result2) < 1e-6);
-                }
-            }
-        }
-
-        [Fact]
-        public void GaussianIsPositiveTest()
-        {
-            for (int trial = 0; trial < 2; trial++)
-            {
-                double xMean = (trial == 0) ? -1 : 1;
-                Gaussian x = Gaussian.FromMeanAndVariance(xMean, 0);
-                Gaussian result2 = IsPositiveOp.XAverageConditional(true, x);
-                //Console.WriteLine("{0}: {1}", x, result2);
-                if (trial == 0)
-                    Assert.True(result2.IsPointMass && result2.Point == 0.0);
-                else
-                    Assert.True(result2.IsUniform());
-                for (int i = 8; i < 30; i++)
-                {
-                    double v = System.Math.Pow(0.1, i);
-                    x = Gaussian.FromMeanAndVariance(xMean, v);
-                    Gaussian result = IsPositiveOp.XAverageConditional(true, x);
-                    //Console.WriteLine("{0}: {1} maxDiff={2}", x, result, result2.MaxDiff(result));
-                    if (trial == 0)
-                    {
-                        Assert.True(MMath.AbsDiff(result2.GetMean(), result.GetMean()) < 1e-6);
-                    }
-                    else
-                        Assert.True(result2.MaxDiff(result) < 1e-6);
-                }
-            }
-            for (int i = 0; i < 20; i++)
-            {
-                Assert.True(IsPositiveOp.XAverageConditional(true, Gaussian.FromNatural(-System.Math.Pow(10, i), 0.1)).IsProper());
-            }
-            Assert.True(IsPositiveOp.XAverageConditional(true, Gaussian.FromNatural(-2.3287253734154412E+107, 0.090258824802119691)).IsProper());
-            Assert.True(IsPositiveOp.XAverageConditional(false, Gaussian.FromNatural(2.3287253734154412E+107, 0.090258824802119691)).IsProper());
-            //Assert.True(IsPositiveOp.XAverageConditional(Bernoulli.FromLogOdds(-4e-16), new Gaussian(490, 1.488e+06)).IsProper());
-            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(-2.03e+09, 5.348e+09)).IsProper());
-
-            Gaussian uniform = new Gaussian();
-            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(Gaussian.FromMeanAndVariance(0, 1)).LogOdds);
-            Assert.True(MMath.AbsDiff(MMath.NormalCdfLogit(2.0 / System.Math.Sqrt(3)), IsPositiveOp.IsPositiveAverageConditional(Gaussian.FromMeanAndVariance(2, 3)).LogOdds, 1e-10) <
-                          1e-10);
-            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(uniform).LogOdds);
-            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(Gaussian.PointMass(0.0)).GetProbTrue());
-
-            Bernoulli isPositiveDist = new Bernoulli();
-            Gaussian xDist = new Gaussian(0, 1);
-            Gaussian xMsg = new Gaussian();
-            Assert.True(IsPositiveOp.XAverageConditional(isPositiveDist, xDist).Equals(uniform));
-            isPositiveDist = new Bernoulli(0);
-            Gaussian xBelief = new Gaussian();
-            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
-            xBelief.SetToProduct(xMsg, xDist);
-            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(-0.7979, 0.3634)) < 1e-1);
-            isPositiveDist = new Bernoulli(1);
-            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
-            xBelief.SetToProduct(xMsg, xDist);
-            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(0.7979, 0.3634)) < 1e-1);
-            isPositiveDist = new Bernoulli(0.6);
-            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
-            xBelief.SetToProduct(xMsg, xDist);
-            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(0.15958, 0.97454)) < 1e-1);
-            Assert.True(IsPositiveOp.XAverageConditional(isPositiveDist, uniform).Equals(uniform));
-
-            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(127, 11)).Equals(uniform));
-            Assert.True(IsPositiveOp.XAverageConditional(false, new Gaussian(-127, 11)).Equals(uniform));
-            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(-1e5, 10)).IsProper());
-            Assert.True(IsPositiveOp.XAverageConditional(false, new Gaussian(1e5, 10)).IsProper());
         }
 
         [Fact]
@@ -3302,6 +2116,19 @@ weight * (tau + alphaX) + alphaX
         }
 
         [Fact]
+        public void IsBetweenGaussian_WithImprobableInputs_Throws()
+        {
+            Bernoulli isBetween = new Bernoulli(1);
+            Gaussian X = Gaussian.FromMeanAndVariance(double.MinValue / 2, 1);
+            Gaussian lowerBound = Gaussian.FromMeanAndVariance(double.MaxValue / 2, 1);
+            Gaussian upperBound = Gaussian.FromNatural(0, 4);
+            Assert.Throws<AllZeroException>(() =>
+                IsBetweenGaussianOp.XAverageConditional_Slow(isBetween, X, lowerBound, upperBound));
+            Assert.Throws<AllZeroException>(() =>
+                IsBetweenGaussianOp.LowerBoundAverageConditional_Slow(isBetween, X, lowerBound, upperBound));
+        }
+
+        [Fact]
         public void GaussianIsBetweenTest2()
         {
             Bernoulli isBetween = new Bernoulli(1);
@@ -3461,253 +2288,6 @@ weight * (tau + alphaX) + alphaX
             //Assert.True(Upost.MaxDiff(IsPositiveOp.XAverageConditional(true,upperBound)) < 1e-3);
         }
 
-        internal void ProductGaussianGammaVmpOpTest()
-        {
-            double prec = 1;
-            double a = 2;
-            Gamma B = Gamma.FromShapeAndRate(1e-3, 1e-3);
-            //Gamma B = Gamma.FromShapeAndRate(1, 1);
-            for (int i = 0; i < 10; i++)
-            {
-                prec = (i + 1) * 4.0e-15;
-                Gamma result = ProductGaussianGammaVmpOp.BAverageLogarithm(Gaussian.FromNatural(0.5, prec), Gaussian.PointMass(a), B, Gamma.Uniform());
-                Gamma result2 = ProductGaussianGammaVmpOp.BAverageLogarithm(Gaussian.FromNatural(0.5, prec), Gaussian.PointMass(a), B, Gamma.Uniform());
-                Console.WriteLine("{0} {1}", result, result2);
-                //Console.WriteLine("rate = {0}", result.Rate);
-                // rate = -0.5*A
-            }
-        }
-
-        [Fact]
-        public void ProductOpTest()
-        {
-            Assert.True(GaussianProductVmpOp.ProductAverageLogarithm(
-              2.0,
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5)) < 1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(
-              2.0,
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5)) < 1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(new Gaussian(0, 1),
-              Gaussian.PointMass(2.0),
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5)) < 1e-8);
-            Assert.True(GaussianProductVmpOp.ProductAverageLogarithm(
-              0.0,
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.PointMass(0.0)) < 1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(
-              0.0,
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.PointMass(0.0)) < 1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(new Gaussian(0, 1),
-              Gaussian.PointMass(0.0),
-              Gaussian.FromMeanAndVariance(3.0, 5.0)).MaxDiff(Gaussian.PointMass(0.0)) < 1e-8);
-
-            Assert.True(GaussianProductVmpOp.ProductAverageLogarithm(
-              Gaussian.FromMeanAndVariance(2, 4),
-              Gaussian.FromMeanAndVariance(3, 5)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5 + 3 * 3 * 4 + 2 * 2 * 5)) < 1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(Gaussian.Uniform(),
-              Gaussian.FromMeanAndVariance(2, 4),
-                                                                            Gaussian.FromMeanAndVariance(3, 5)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5 + 3 * 3 * 4 + 2 * 2 * 5)) <
-                                1e-8);
-            Assert.True(GaussianProductOp.ProductAverageConditional(Gaussian.FromMeanAndVariance(0, 1e16),
-              Gaussian.FromMeanAndVariance(2, 4),
-                                                                            Gaussian.FromMeanAndVariance(3, 5)).MaxDiff(Gaussian.FromMeanAndVariance(2 * 3, 4 * 5 + 3 * 3 * 4 + 2 * 2 * 5)) <
-                                1e-4);
-
-            Assert.True(GaussianProductOp.AAverageConditional(6.0, 2.0)
-              .MaxDiff(Gaussian.PointMass(6.0 / 2.0)) < 1e-8);
-            Assert.True(GaussianProductOp.AAverageConditional(6.0, new Gaussian(1, 3), Gaussian.PointMass(2.0))
-              .MaxDiff(Gaussian.PointMass(6.0 / 2.0)) < 1e-8);
-            Assert.True(GaussianProductOp.AAverageConditional(0.0, 0.0).IsUniform());
-            Assert.True(GaussianProductOp.AAverageConditional(Gaussian.Uniform(), 2.0).IsUniform());
-            Assert.True(GaussianProductOp.AAverageConditional(Gaussian.Uniform(), new Gaussian(1, 3), Gaussian.PointMass(2.0)).IsUniform());
-            Assert.True(GaussianProductOp.AAverageConditional(Gaussian.Uniform(), new Gaussian(1, 3), new Gaussian(2, 4)).IsUniform());
-
-            Gaussian aPrior = Gaussian.FromMeanAndVariance(0.0, 1000.0);
-            Assert.True((GaussianProductOp.AAverageConditional(
-              Gaussian.FromMeanAndVariance(10.0, 1.0),
-              aPrior,
-              Gaussian.FromMeanAndVariance(5.0, 1.0)) * aPrior).MaxDiff(
-              Gaussian.FromMeanAndVariance(2.208041421368822, 0.424566765678152)) < 1e-4);
-
-            Gaussian g = new Gaussian(0, 1);
-            Assert.True(GaussianProductOp.AAverageConditional(g, 0.0).IsUniform());
-            Assert.True(GaussianProductOp.AAverageConditional(0.0, 0.0).IsUniform());
-            Assert.True(GaussianProductVmpOp.AAverageLogarithm(g, 0.0).IsUniform());
-            Assert.True(Gaussian.PointMass(3.0).MaxDiff(GaussianProductVmpOp.AAverageLogarithm(6.0, 2.0)) < 1e-10);
-            Assert.True(GaussianProductVmpOp.AAverageLogarithm(0.0, 0.0).IsUniform());
-            try
-            {
-                Assert.True(GaussianProductVmpOp.AAverageLogarithm(6.0, g).IsUniform());
-                Assert.True(false, "Did not throw NotSupportedException");
-            }
-            catch (NotSupportedException)
-            {
-            }
-            try
-            {
-                g = GaussianProductOp.AAverageConditional(12.0, 0.0);
-                Assert.True(false, "Did not throw AllZeroException");
-            }
-            catch (AllZeroException)
-            {
-            }
-            try
-            {
-                g = GaussianProductVmpOp.AAverageLogarithm(12.0, 0.0);
-                Assert.True(false, "Did not throw AllZeroException");
-            }
-            catch (AllZeroException)
-            {
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "ModifiesGlobals")]
-        public void GaussianProductOp_APointMassTest()
-        {
-            using (TestUtils.TemporarilyAllowGaussianImproperMessages)
-            {
-                Gaussian Product = Gaussian.FromMeanAndVariance(1.3, 0.1);
-                Gaussian B = Gaussian.FromMeanAndVariance(1.24, 0.04);
-                GaussianProductOp_APointMass(1, Product, B);
-
-                Product = Gaussian.FromMeanAndVariance(10, 1);
-                B = Gaussian.FromMeanAndVariance(5, 1);
-                GaussianProductOp_APointMass(2, Product, B);
-
-                Product = Gaussian.FromNatural(1, 0);
-                GaussianProductOp_APointMass(2, Product, B);
-            }
-        }
-
-        private void GaussianProductOp_APointMass(double aMean, Gaussian Product, Gaussian B)
-        {
-            bool isProper = Product.IsProper();
-            Gaussian A = Gaussian.PointMass(aMean);
-            Gaussian result = GaussianProductOp.AAverageConditional(Product, A, B);
-            Console.WriteLine("{0}: {1}", A, result);
-            Gaussian result2 = isProper ? GaussianProductOp_Slow.AAverageConditional(Product, A, B) : result;
-            Console.WriteLine("{0}: {1}", A, result2);
-            Assert.True(result.MaxDiff(result2) < 1e-6);
-            var Amsg = InnerProductOp_PointB.BAverageConditional(Product, DenseVector.FromArray(B.GetMean()), new PositiveDefiniteMatrix(new double[,] { { B.GetVariance() } }), VectorGaussian.PointMass(aMean), VectorGaussian.Uniform(1));
-            //Console.WriteLine("{0}: {1}", A, Amsg);
-            Assert.True(result.MaxDiff(Amsg.GetMarginal(0)) < 1e-6);
-            double prevDiff = double.PositiveInfinity;
-            for (int i = 3; i < 40; i++)
-            {
-                double v = System.Math.Pow(0.1, i);
-                A = Gaussian.FromMeanAndVariance(aMean, v);
-                result2 = isProper ? GaussianProductOp.AAverageConditional(Product, A, B) : result;
-                double diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} diff={2}", A, result2, diff.ToString("g4"));
-                //Assert.True(diff <= prevDiff || diff < 1e-6);
-                result2 = isProper ? GaussianProductOp_Slow.AAverageConditional(Product, A, B) : result;
-                diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} diff={2}", A, result2, diff.ToString("g4"));
-                Assert.True(diff <= prevDiff || diff < 1e-6);
-                prevDiff = diff;
-            }
-        }
-
-        [Fact]
-        [Trait("Category", "OpenBug")]
-        public void GaussianProductOp_ProductPointMassTest()
-        {
-            Gaussian A = new Gaussian(1, 2);
-            Gaussian B = new Gaussian(3, 4);
-            Gaussian pointMass = Gaussian.PointMass(4);
-            Gaussian to_pointMass = GaussianProductOp.ProductAverageConditional(pointMass, A, B);
-            double prevDiff = double.PositiveInfinity;
-            for (int i = 0; i < 100; i++)
-            {
-                Gaussian Product = Gaussian.FromMeanAndVariance(4, System.Math.Pow(10, -i));
-                Gaussian to_product = GaussianProductOp.ProductAverageConditional(Product, A, B);
-                double evidence = GaussianProductOp.LogEvidenceRatio(Product, A, B, to_product);
-                Console.WriteLine($"{Product} {to_product} {evidence}");
-                double diff = to_product.MaxDiff(to_pointMass);
-                Assert.True(diff <= prevDiff || diff < 1e-6);
-                prevDiff = diff;
-            }
-        }
-
-        [Fact]
-        public void ProductOpTest3()
-        {
-            Gaussian Product = new Gaussian(3.207, 2.222e-06);
-            Gaussian A = new Gaussian(2.854e-06, 1.879e-05);
-            Gaussian B = new Gaussian(0, 1);
-            Gaussian result = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-            Console.WriteLine(result);
-            Assert.False(double.IsNaN(result.Precision));
-
-            Product = Gaussian.FromNatural(2, 1);
-            A = Gaussian.FromNatural(0, 3);
-            B = Gaussian.FromNatural(0, 1);
-            result = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-            Console.WriteLine("{0}: {1}", Product, result);
-
-            Product = Gaussian.FromNatural(129146.60457039363, 320623.20967711863);
-            A = Gaussian.FromNatural(-0.900376203577801, 0.00000001);
-            B = Gaussian.FromNatural(0, 1);
-            result = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-            Console.WriteLine("{0}: {1}", Product, result);
-
-            Assert.True(GaussianProductOp_Slow.ProductAverageConditional(
-              Gaussian.FromMeanAndVariance(0.0, 1000.0),
-              Gaussian.FromMeanAndVariance(2.0, 3.0),
-              Gaussian.FromMeanAndVariance(5.0, 1.0)).MaxDiff(
-              Gaussian.FromMeanAndVariance(9.911, 79.2)
-              // Gaussian.FromMeanAndVariance(12.110396063215639,3.191559311624262e+002)
-              ) < 1e-4);
-
-            A = new Gaussian(2, 3);
-            B = new Gaussian(4, 5);
-            Product = Gaussian.PointMass(2);
-            result = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-            Console.WriteLine("{0}: {1}", Product, result);
-            double prevDiff = double.PositiveInfinity;
-            for (int i = 3; i < 40; i++)
-            {
-                double v = System.Math.Pow(0.1, i);
-                Product = Gaussian.FromMeanAndVariance(2, v);
-                Gaussian result2 = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-                double diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} diff={2}", Product, result2, diff.ToString("g4"));
-                Assert.True(diff <= prevDiff || diff < 1e-6);
-                prevDiff = diff;
-            }
-
-            Product = Gaussian.Uniform();
-            result = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-            Console.WriteLine("{0}: {1}", Product, result);
-            prevDiff = double.PositiveInfinity;
-            for (int i = 3; i < 40; i++)
-            {
-                double v = System.Math.Pow(10, i);
-                Product = Gaussian.FromMeanAndVariance(2, v);
-                Gaussian result2 = GaussianProductOp_Slow.ProductAverageConditional(Product, A, B);
-                double diff = result.MaxDiff(result2);
-                Console.WriteLine("{0}: {1} diff={2}", Product, result2, diff.ToString("g4"));
-                Assert.True(diff <= prevDiff || diff < 1e-6);
-                prevDiff = diff;
-            }
-        }
-
-        [Fact]
-        public void LogisticProposalDistribution()
-        {
-            double[] TrueCounts = { 2, 20, 200, 2000, 20000, 2, 20, 200, 2000, 20000 };
-            double[] FalseCounts = { 2, 200, 20000, 200, 20, 200, 2000, 2, 2000, 20 };
-            double[] Means = { .5, 1, 2, 4, 8, 16, 32, 0, -2, -20 };
-            double[] Variances = { 1, 2, 4, 8, .1, .0001, .01, .000001, 0.000001, 0.001 };
-            for (int i = 0; i < 10; i++)
-            {
-                Beta b = new Beta();
-                b.TrueCount = TrueCounts[i];
-                b.FalseCount = FalseCounts[i];
-                Gaussian g = Gaussian.FromMeanAndVariance(Means[i], Variances[i]);
-                Gaussian gProposal = GaussianBetaProductOp.LogisticProposalDistribution(b, g);
-            }
-        }
 
         /// <summary>
         /// Tests for the message operators for <see cref="Factor.CountTrue"/>.
@@ -3777,30 +2357,477 @@ weight * (tau + alphaX) + alphaX
                 CountTrueOp.ArrayAverageConditional(array, new Discrete(1.0), forwardPassBuffer, resultArray);
             }
         }
-    }
 
-    public class ImportanceSampler
-    {
-        public delegate double Sampler();
-        private Vector samples, weights;
-        public ImportanceSampler(int N, Sampler sampler, Converter<double, double> weightFunction)
+        [Fact]
+        public void VariablePointOp_RpropGammaTest()
         {
-            samples = Vector.Zero(N);
-            weights = Vector.Zero(N);
-            for (int i = 0; i < N; i++)
+            using (TestUtils.TemporarilyUseMeanPointGamma)
             {
-                samples[i] = sampler();
-                weights[i] = weightFunction(samples[i]);
+                var buffer = VariablePointOp_RpropGamma.Buffer0Init();
+                Gamma g = Gamma.FromShapeAndRate(0.6, 1);
+                Gamma to_marginal = Gamma.PointMass(1);
+                for (int i = 0; i < 1000; i++)
+                {
+                    buffer = VariablePointOp_RpropGamma.Buffer0(g, g, to_marginal, buffer);
+                    //Console.WriteLine(buffer.nextPoint);
+                    to_marginal = Gamma.Uniform();
+                }
             }
         }
 
-        public double GetExpectation(Converter<double, double> h)
+        [Fact]
+        public void ArrayFromVectorOpTest()
         {
-            var temp = Vector.Zero(samples.Count);
-            temp.SetToFunction(samples, h);
-            temp.SetToProduct(temp, weights);
-            return temp.Sum() / weights.Sum();
+            bool verbose = false;
+            PositiveDefiniteMatrix A = new PositiveDefiniteMatrix(new double[,] {
+                { 4.008640513161180,  1.303104352135630, - 2.696380025254830, - 2.728465435435790 },
+                { 1.303104352135630,  4.024136989099960, - 2.681070246787840, - 2.713155656968810 },
+                { -2.696380025254830, - 2.681070246787840, 4.136120496920130,  1.403451295855420 },
+                { -2.728465435435790, - 2.713155656968810, 1.403451295855420,  4.063123100392480 }
+            });
+            Vector arrayVariance = Vector.FromArray(1, 0, 3, 4);
+            Vector arrayMean = Vector.FromArray(1, 2, 3, 4);
+
+            VectorGaussianMoments vg = new VectorGaussianMoments(Vector.FromArray(6, 5, 4, 3), A);
+            Gaussian[] array = Util.ArrayInit(arrayMean.Count, i => new Gaussian(arrayMean[i], arrayVariance[i]));
+            var result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
+            Vector varianceExpected = Vector.FromArray(0.699231932932321, 0, 0.946948669226297, 0.926496676481940);
+            Vector varianceActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetVariance()));
+            if (verbose) Console.WriteLine($"variance = {varianceActual} should be {varianceExpected}");
+            Assert.True(varianceExpected.MaxDiff(varianceActual) < 1e-4);
+            Vector meanExpected = Vector.FromArray(2.640276200841019, 2.000000014880260, 6.527941507328482, 6.908179339051594);
+            Vector meanActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetMean()));
+            if (verbose) Console.WriteLine($"mean = {meanActual} should be {meanExpected}");
+            Assert.True(meanExpected.MaxDiff(meanActual) < 1e-4);
+
+            arrayVariance.SetTo(Vector.FromArray(1, double.PositiveInfinity, 3, 4));
+            array = Util.ArrayInit(arrayMean.Count, i => new Gaussian(arrayMean[i], arrayVariance[i]));
+            result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
+            varianceExpected = Vector.FromArray(0.703202829692760, 2.371534574978036, 1.416576208767429, 1.566923316764467);
+            varianceActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetVariance()));
+            if (verbose) Console.WriteLine($"variance = {varianceActual} should be {varianceExpected}");
+            Assert.True(varianceExpected.MaxDiff(varianceActual) < 1e-4);
+            meanExpected = Vector.FromArray(2.495869677904531, 5.528904975989440, 4.957577684071078, 5.074347889058121);
+            meanActual = Vector.FromArray(Util.ArrayInit(result.Length, i => (result[i] * array[i]).GetMean()));
+            if (verbose) Console.WriteLine($"mean = {meanActual} should be {meanExpected}");
+            Assert.True(meanExpected.MaxDiff(meanActual) < 1e-4);
+
+            bool testImproper = false;
+            if (testImproper)
+            {
+                array = new Gaussian[]
+                {
+                new Gaussian(0.5069, 27.43),
+                new Gaussian(0.649, 38.94),
+                Gaussian.FromNatural(-0.5753, 0),
+                new Gaussian(0.1064, 10.15)
+                };
+                PositiveDefiniteMatrix B = new PositiveDefiniteMatrix(new double[,]
+                {
+                { 0.7158,  -0.1356, -0.2979, -0.2993 },
+                { -0.1356, 0.7294,  -0.2975, -0.2989 },
+                { -0.2979, -0.2975, 0.7625,  -0.1337 },
+                { -0.2993, -0.2989, -0.1337, 0.7203 },
+                });
+                vg = new VectorGaussianMoments(Vector.FromArray(0.3042, 0.4795, -0.2114, -0.5748), B);
+                result = ArrayFromVectorOp.ArrayAverageConditional(array, vg, new Gaussian[array.Length]);
+                Gaussian[] resultExpected = new Gaussian[] {
+                    new Gaussian(0.4589, 0.707),
+                    new Gaussian(0.6338, 0.7204),
+                    new Gaussian(-0.2236, 0.7553),
+                    new Gaussian(-0.4982, 0.7148),
+                };
+                if (verbose) Console.WriteLine(StringUtil.ArrayToString(result));
+                for (int i = 0; i < result.Length; i++)
+                {
+                    Assert.True(resultExpected[i].MaxDiff(result[i]) < 1e-4);
+                }
+            }
         }
 
+        internal void VectorFromArrayOp_SpeedTest()
+        {
+            int dim = 20;
+            int n = 100000;
+            VectorGaussian vector = new VectorGaussian(dim);
+            Matrix random = new Matrix(dim, dim);
+            for (int i = 0; i < dim; i++)
+            {
+                vector.MeanTimesPrecision[i] = Rand.Double();
+                for (int j = 0; j < dim; j++)
+                {
+                    random[i, j] = Rand.Double();
+                }
+            }
+            vector.Precision.SetToOuter(random);
+            Gaussian[] array = Util.ArrayInit(dim, i => Gaussian.FromMeanAndVariance(i, i));
+            Gaussian[] result = new Gaussian[dim];
+            for (int i = 0; i < n; i++)
+            {
+                VectorFromArrayOp.ArrayAverageConditional(vector, array, result);
+            }
+        }
+
+        [Fact]
+        public void VectorFromArrayOp_PointMassTest()
+        {
+            VectorFromArrayOp_HandlesPointMass(false);
+            VectorFromArrayOp_HandlesPointMass(true);
+        }
+
+        private void VectorFromArrayOp_HandlesPointMass(bool partial)
+        {
+            int dim = 2;
+            VectorGaussian to_vector = new VectorGaussian(dim);
+            VectorGaussian vector = VectorGaussian.FromNatural(Vector.FromArray(2.0, 3.0), new PositiveDefiniteMatrix(new double[,] { { 5.0, 1.0 }, { 1.0, 6.0 } }));
+            GaussianArray expected = null;
+            double lastError = double.PositiveInfinity;
+            for (int i = 0; i < 10; i++)
+            {
+                double variance = System.Math.Exp(-i);
+                if (i == 0)
+                    variance = 0;
+                Gaussian[] array = Util.ArrayInit(dim, j => new Gaussian(j, (j > 0 && partial) ? j : variance));
+                to_vector = VectorFromArrayOp.VectorAverageConditional(array, to_vector);
+                var to_array = new GaussianArray(dim);
+                to_array = VectorFromArrayOp.ArrayAverageConditional(vector, array, to_array);
+                if (i == 0)
+                    expected = to_array;
+                else
+                {
+                    double error = expected.MaxDiff(to_array);
+                    Assert.True(error < lastError);
+                    lastError = error;
+                }
+                ////Console.WriteLine(to_array);
+            }
+        }
+
+
+        [Fact]
+        public void MatrixTimesVectorOpTest()
+        {
+            Matrix[] ms = new Matrix[3];
+            ms[0] = new Matrix(new double[,] { { 0.036, -0.036, 0 }, { 0.036, 0, -0.036 } });
+            ms[1] = new Matrix(new double[,] { { -0.036, 0.036, 0 }, { 0, 0.036, -0.036 } });
+            ms[2] = new Matrix(new double[,] { { -0.036, 0, 0.036 }, { 0, -0.036, 0.036 } });
+            VectorGaussian product = new VectorGaussian(Vector.FromArray(new double[] { 1, 1 }), PositiveDefiniteMatrix.IdentityScaledBy(2, 0.01));
+            Matrix A = ms[0];
+            VectorGaussian result = MatrixVectorProductOp.BAverageConditional(product, A, new VectorGaussian(3));
+            Assert.False(result.Precision.IsPositiveDefinite());
+
+            Vector BMean = Vector.FromArray(1, 2, 3);
+            PositiveDefiniteMatrix BVariance = new PositiveDefiniteMatrix(new double[,]
+            {
+                { 3,2,1 },
+                { 2,3,2 },
+                { 1,2,3 }
+            });
+            GaussianArray2D to_A = new GaussianArray2D(A.Rows, A.Cols);
+            MatrixVectorProductOp.AAverageConditional(product, GaussianArray2D.PointMass(A.ToArray()), BMean, BVariance, to_A);
+            double[,] dExpected = new double[,]
+            {
+                { 542.014350893474, -34.0730521080406, -261.179402050882 },
+                { 449.735198014871, -31.4263853641523, -215.272881589687 },
+            };
+            for (int i = 0; i < to_A.GetLength(0); i++)
+            {
+                for (int j = 0; j < to_A.GetLength(1); j++)
+                {
+                    Gaussian dist = to_A[i, j];
+                    double dlogp, ddlogp;
+                    dist.GetDerivatives(A[i, j], out dlogp, out ddlogp);
+                    Assert.True(System.Math.Abs(dExpected[i, j] - dlogp) < 1e-8);
+                }
+            }
+        }
+
+        [Fact]
+        public void SumOpTest()
+        {
+            Gaussian sum_F = new Gaussian(0, 1);
+            Gaussian sum_B = Gaussian.FromMeanAndPrecision(0, 1e-310);
+            GaussianArray array = new GaussianArray(2, i => new Gaussian(0, 1));
+            var result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
+            Assert.True(!double.IsNaN(result[0].MeanTimesPrecision));
+
+            sum_B = new Gaussian(0, 1);
+            array[0] = Gaussian.FromMeanAndPrecision(0, 0);
+            sum_F = FastSumOp.SumAverageConditional(array);
+            result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
+            Assert.True(result[0].MaxDiff(new Gaussian(0, 2)) < 1e-10);
+            Assert.True(result[1].IsUniform());
+            for (int i = 0; i < 1030; i++)
+            {
+                array[0] = Gaussian.FromMeanAndPrecision(0, System.Math.Pow(2, -i));
+                sum_F = FastSumOp.SumAverageConditional(array);
+                result = FastSumOp.ArrayAverageConditional(sum_B, sum_F, array, new GaussianArray(array.Count));
+                Assert.True(result[0].MaxDiff(new Gaussian(0, 2)) < 1e-10);
+            }
+        }
+
+
+        [Fact]
+        public void LogisticOpTest()
+        {
+            for (int trial = 0; trial < 2; trial++)
+            {
+                double xMean = (trial == 0) ? -1 : 1;
+                Gaussian x = Gaussian.FromMeanAndVariance(xMean, 0);
+                Gaussian result2 = BernoulliFromLogOddsOp.LogOddsAverageConditional(true, x);
+                Console.WriteLine("{0}: {1}", x, result2);
+                for (int i = 8; i < 30; i++)
+                {
+                    double v = System.Math.Pow(0.1, i);
+                    x = Gaussian.FromMeanAndVariance(xMean, v);
+                    Gaussian result = BernoulliFromLogOddsOp.LogOddsAverageConditional(true, x);
+                    Console.WriteLine("{0}: {1} maxDiff={2}", x, result, result2.MaxDiff(result));
+                    Assert.True(result2.MaxDiff(result) < 1e-6);
+                }
+            }
+
+            for (int i = 0; i < 10; i++)
+            {
+                Gaussian falseMsg = LogisticOp.FalseMsg(new Beta(0.2, 1.8), new Gaussian(0, 97.0 * (i + 1)), new Gaussian());
+                Console.WriteLine(falseMsg);
+            }
+            Gaussian toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(-4662, 1314));
+            Console.WriteLine(toLogOdds);
+            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(2249, 2.5));
+            Console.WriteLine(toLogOdds);
+            Gaussian logOdds = new Gaussian(100, 100);
+            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, logOdds);
+            Console.WriteLine(toLogOdds * logOdds);
+            // test m =approx 1.5*v for increasing v
+            for (int i = 0; i < 10; i++)
+            {
+                double v = System.Math.Pow(2, i + 5);
+                Gaussian g = new Gaussian(37 + 1.5 * v, v);
+                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, g);
+                Console.WriteLine("{0}: {1}", g, toLogOdds);
+                Gaussian actualPost = toLogOdds * g;
+                Gaussian expectedPost = new Gaussian(0, 1);
+                if (i == 0)
+                {
+                    // for Gaussian(85,32)
+                    expectedPost = new Gaussian(53, 32);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+                else if (i == 1)
+                {
+                    // for Gaussian(133,64)
+                    expectedPost = new Gaussian(69, 64);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+                else if (i == 4)
+                {
+                    // for Gaussian(805,512) the posterior should be (293, 511.81)
+                    expectedPost = new Gaussian(293, 511.81);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+                else if (i == 5)
+                {
+                    // for Gaussian(1573,1024)
+                    expectedPost = new Gaussian(549, 1023.9);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+            }
+            // test m =approx v for increasing v
+            for (int i = 0; i < 10; i++)
+            {
+                double v = System.Math.Pow(2, i + 5);
+                Gaussian g = new Gaussian(v - 1, v);
+                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, g);
+                Console.WriteLine("{0}: {1}", g, toLogOdds);
+                Gaussian actualPost = toLogOdds * g;
+                Gaussian expectedPost = new Gaussian(0, 1);
+                if (i == 0)
+                {
+                    // for Gaussian(31,32)
+                    expectedPost = new Gaussian(3.8979, 12.4725);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+                else if (i == 6)
+                {
+                    // for Gaussian(2047,2048)
+                    expectedPost = new Gaussian(35.7156, 736.2395);
+                    Assert.True(expectedPost.MaxDiff(actualPost) < 1e-3);
+                }
+            }
+            for (int i = 0; i < 10; i++)
+            {
+                toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(54.65 / 10 * (i + 1), 8.964));
+                Console.WriteLine(toLogOdds);
+            }
+            toLogOdds = BernoulliFromLogOddsOp.LogOddsAverageConditional(false, new Gaussian(9900, 10000) ^ 0.1);
+            Console.WriteLine(toLogOdds);
+            Gaussian falseMsg2 = LogisticOp.FalseMsg(new Beta(0.9, 0.1), Gaussian.FromNatural(-10.097766458353044, 0.000011644704327819733),
+              Gaussian.FromNatural(-0.0010832099815010626, 0.000010092906656322242));
+            Console.WriteLine(falseMsg2);
+        }
+
+        [Fact]
+        public void BernoulliFromLogOddsTest()
+        {
+            double tolerance = 1e-8;
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(0, 2)), System.Math.Log(0.5)) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(2, 0)), MMath.LogisticLn(2)) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(1, 1)), System.Math.Log(1 - 0.5 * System.Math.Exp(-0.5))) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(10, 10)), System.Math.Log(1 - 0.5 * System.Math.Exp(-5))) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-1, 1)), System.Math.Log(0.303265329856008)) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-5, 3)), System.Math.Log(0.023962216989475)) < tolerance);
+            Assert.True(MMath.AbsDiff(BernoulliFromLogOddsOp.LogAverageFactor(true, new Gaussian(-5, 10)), System.Math.Log(0.084396512538678)) < tolerance);
+        }
+
+
+        [Fact]
+        public void OrTest()
+        {
+            Assert.Equal(1.0, BooleanOrOp.AAverageConditional(true, false).GetProbTrue());
+            Assert.Equal(1.0, BooleanOrOp.OrAverageConditional(Bernoulli.PointMass(true), false).GetProbTrue());
+
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), Bernoulli.PointMass(false)).LogOdds);
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, Bernoulli.PointMass(false)).LogOdds);
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), false).LogOdds);
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, false).LogOdds);
+
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), Bernoulli.PointMass(true)).GetProbTrue());
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, Bernoulli.PointMass(true)).GetProbTrue());
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(Bernoulli.PointMass(false), true).GetProbTrue());
+            Assert.Equal(0.0, BooleanAndOp.AAverageConditional(false, true).GetProbTrue());
+
+            Assert.Throws<AllZeroException>(() =>
+            {
+                BooleanAndOp.AAverageConditional(true, false);
+            });
+            Assert.Throws<AllZeroException>(() =>
+            {
+                BooleanAndOp.AAverageConditional(Bernoulli.PointMass(true), false);
+            });
+            Assert.Throws<AllZeroException>(() =>
+            {
+                BooleanAndOp.AAverageConditional(true, Bernoulli.PointMass(false));
+            });
+            Assert.Throws<AllZeroException>(() =>
+            {
+                BooleanAndOp.AAverageConditional(Bernoulli.PointMass(true), Bernoulli.PointMass(false));
+            });
+        }
+
+        [Fact]
+        public void BernoulliFromDiscreteTest()
+        {
+            double[] probTrue = { 0.4, 0.7 };
+            Discrete indexDist = new Discrete(0.8, 0.2);
+            Bernoulli sampleDist = new Bernoulli(0.8 * 0.4 + 0.2 * 0.7);
+            Assert.True(sampleDist.MaxDiff(BernoulliFromDiscreteOp.SampleAverageConditional(indexDist, probTrue)) < 1e-4);
+            sampleDist.SetProbTrue(0.2);
+            double p = (0.8 * 0.3 + 0.2 * 0.7) / (0.8 * 0.3 + 0.2 * 0.7 + 0.8 * 0.6 + 0.2 * 0.4);
+            Vector probs = indexDist.GetWorkspace();
+            probs[0] = 1 - p;
+            probs[1] = p;
+            indexDist.SetProbs(probs);
+            Assert.True(indexDist.MaxDiff(BernoulliFromDiscreteOp.IndexAverageConditional(sampleDist, probTrue, Discrete.Uniform(indexDist.Dimension))) < 1e-4);
+        }
+
+        [Fact]
+        public void BernoulliFromBooleanTest()
+        {
+            double[] probTrue = { 0.4, 0.7 };
+            Bernoulli choiceDist = new Bernoulli(0.2);
+            Bernoulli sampleDist = new Bernoulli(0.8 * 0.4 + 0.2 * 0.7);
+            Assert.True(sampleDist.MaxDiff(BernoulliFromBooleanArray.SampleAverageConditional(choiceDist, probTrue)) < 1e-4);
+            sampleDist.SetProbTrue(0.2);
+            choiceDist.SetProbTrue((0.8 * 0.3 + 0.2 * 0.7) / (0.8 * 0.3 + 0.2 * 0.7 + 0.8 * 0.6 + 0.2 * 0.4));
+            Assert.True(choiceDist.MaxDiff(BernoulliFromBooleanArray.ChoiceAverageConditional(sampleDist, probTrue)) < 1e-4);
+        }
+
+
+        // test the limit of an incoming message with small precision
+        [Fact]
+        public void IsPositiveGaussianTest2()
+        {
+            for (int trial = 0; trial < 2; trial++)
+            {
+                bool isPositive = (trial == 0);
+                Gaussian x = Gaussian.FromNatural(isPositive ? -2 : 2, 0);
+                Gaussian result = IsPositiveOp.XAverageConditional(isPositive, x);
+                for (int i = 10; i < 20; i++)
+                {
+                    x.Precision = System.Math.Pow(0.1, i);
+                    Gaussian result2 = IsPositiveOp.XAverageConditional(isPositive, x);
+                    //Console.WriteLine($"{x}: {result2} maxDiff={result.MaxDiff(result2)}");
+                    Assert.True(result.MaxDiff(result2) < 1e-6);
+                }
+            }
+        }
+
+        [Fact]
+        public void IsPositiveGaussianTest()
+        {
+            for (int trial = 0; trial < 2; trial++)
+            {
+                double xMean = (trial == 0) ? -1 : 1;
+                Gaussian x = Gaussian.FromMeanAndVariance(xMean, 0);
+                Gaussian result2 = IsPositiveOp.XAverageConditional(true, x);
+                //Console.WriteLine("{0}: {1}", x, result2);
+                if (trial == 0)
+                    Assert.True(result2.IsPointMass && result2.Point == 0.0);
+                else
+                    Assert.True(result2.IsUniform());
+                for (int i = 8; i < 30; i++)
+                {
+                    double v = System.Math.Pow(0.1, i);
+                    x = Gaussian.FromMeanAndVariance(xMean, v);
+                    Gaussian result = IsPositiveOp.XAverageConditional(true, x);
+                    //Console.WriteLine("{0}: {1} maxDiff={2}", x, result, result2.MaxDiff(result));
+                    if (trial == 0)
+                    {
+                        Assert.True(MMath.AbsDiff(result2.GetMean(), result.GetMean()) < 1e-6);
+                    }
+                    else
+                        Assert.True(result2.MaxDiff(result) < 1e-6);
+                }
+            }
+            for (int i = 0; i < 20; i++)
+            {
+                Assert.True(IsPositiveOp.XAverageConditional(true, Gaussian.FromNatural(-System.Math.Pow(10, i), 0.1)).IsProper());
+            }
+            Assert.True(IsPositiveOp.XAverageConditional(true, Gaussian.FromNatural(-2.3287253734154412E+107, 0.090258824802119691)).IsProper());
+            Assert.True(IsPositiveOp.XAverageConditional(false, Gaussian.FromNatural(2.3287253734154412E+107, 0.090258824802119691)).IsProper());
+            //Assert.True(IsPositiveOp.XAverageConditional(Bernoulli.FromLogOdds(-4e-16), new Gaussian(490, 1.488e+06)).IsProper());
+            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(-2.03e+09, 5.348e+09)).IsProper());
+
+            Gaussian uniform = new Gaussian();
+            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(Gaussian.FromMeanAndVariance(0, 1)).LogOdds);
+            Assert.True(MMath.AbsDiff(MMath.NormalCdfLogit(2.0 / System.Math.Sqrt(3)), IsPositiveOp.IsPositiveAverageConditional(Gaussian.FromMeanAndVariance(2, 3)).LogOdds, 1e-10) <
+                          1e-10);
+            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(uniform).LogOdds);
+            Assert.Equal(0.0, IsPositiveOp.IsPositiveAverageConditional(Gaussian.PointMass(0.0)).GetProbTrue());
+
+            Bernoulli isPositiveDist = new Bernoulli();
+            Gaussian xDist = new Gaussian(0, 1);
+            Gaussian xMsg = new Gaussian();
+            Assert.True(IsPositiveOp.XAverageConditional(isPositiveDist, xDist).Equals(uniform));
+            isPositiveDist = new Bernoulli(0);
+            Gaussian xBelief = new Gaussian();
+            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
+            xBelief.SetToProduct(xMsg, xDist);
+            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(-0.7979, 0.3634)) < 1e-1);
+            isPositiveDist = new Bernoulli(1);
+            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
+            xBelief.SetToProduct(xMsg, xDist);
+            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(0.7979, 0.3634)) < 1e-1);
+            isPositiveDist = new Bernoulli(0.6);
+            xMsg = IsPositiveOp.XAverageConditional(isPositiveDist, xDist);
+            xBelief.SetToProduct(xMsg, xDist);
+            Assert.True(xBelief.MaxDiff(Gaussian.FromMeanAndVariance(0.15958, 0.97454)) < 1e-1);
+            Assert.True(IsPositiveOp.XAverageConditional(isPositiveDist, uniform).Equals(uniform));
+
+            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(127, 11)).Equals(uniform));
+            Assert.True(IsPositiveOp.XAverageConditional(false, new Gaussian(-127, 11)).Equals(uniform));
+            Assert.True(IsPositiveOp.XAverageConditional(true, new Gaussian(-1e5, 10)).IsProper());
+            Assert.True(IsPositiveOp.XAverageConditional(false, new Gaussian(1e5, 10)).IsProper());
+        }
     }
 }
